Load Argo global index files

The argo_mirror() points to a directory containing millions of NetCDF files in which Argo data is organized. These files are indexed by category in the root directory of the mirror. The argo_global_*() functions provide access to these as data frames.

argo_global_meta(download = NULL, quiet = FALSE)

argo_global_prof(download = NULL, quiet = FALSE)

argo_global_tech(download = NULL, quiet = FALSE)

argo_global_traj(download = NULL, quiet = FALSE)

argo_global_bio_traj(download = NULL, quiet = FALSE)

argo_global_bio_prof(download = NULL, quiet = FALSE)

argo_global_synthetic_prof(download = NULL, quiet = FALSE)

Arguments

download

A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet

Use FALSE to show which files are downloaded and for more verbose error messages.

Value

argo_global_meta(): a tibble::tibble() of the Argo meta file index

argo_global_tech(): a tibble::tibble() of the Argo tech file index

argo_global_traj(): a tibble::tibble() of the Argo trajectory file index

argo_global_bio_traj(): a tibble::tibble() of the Argo meta biogeochemical trajectory index

argo_global_prof(): a tibble::tibble() of the Argo profile file index

argo_global_bio_prof(): a tibble::tibble() of the Argo biogeochemical profile index

argo_global_synthetic_prof(): a tibble::tibble() of the Argo biogeochemical synthetic profile index

Examples

with_argo_example_cache({
  argo_global_meta()
})
#> Loading argo_global_meta()
#> # A tibble: 16,062 × 4
#>    file                     profiler_type institution date_update        
#>    <chr>                            <dbl> <chr>       <dttm>             
#>  1 aoml/13857/13857_meta.nc           845 AO          2018-10-11 20:00:14
#>  2 aoml/13858/13858_meta.nc           845 AO          2018-10-11 20:00:15
#>  3 aoml/13859/13859_meta.nc           845 AO          2018-10-11 20:00:25
#>  4 aoml/15819/15819_meta.nc           845 AO          2018-10-11 20:00:16
#>  5 aoml/15820/15820_meta.nc           845 AO          2018-10-11 20:00:18
#>  6 aoml/15851/15851_meta.nc           845 AO          2018-10-11 20:00:26
#>  7 aoml/15852/15852_meta.nc           845 AO          2018-10-11 20:00:28
#>  8 aoml/15853/15853_meta.nc           845 AO          2018-10-11 20:00:29
#>  9 aoml/15854/15854_meta.nc           845 AO          2018-10-11 20:00:30
#> 10 aoml/15855/15855_meta.nc           845 AO          2018-10-11 20:00:34
#> # … with 16,052 more rows

with_argo_example_cache({
  argo_global_tech()
})
#> Loading argo_global_tech()
#> # A tibble: 15,610 × 3
#>    file                     institution date_update        
#>    <chr>                    <chr>       <dttm>             
#>  1 aoml/13857/13857_tech.nc AO          2018-10-11 20:00:14
#>  2 aoml/13858/13858_tech.nc AO          2018-10-11 20:00:15
#>  3 aoml/13859/13859_tech.nc AO          2018-10-11 20:00:25
#>  4 aoml/15819/15819_tech.nc AO          2018-10-11 20:00:16
#>  5 aoml/15820/15820_tech.nc AO          2018-10-11 20:00:18
#>  6 aoml/15851/15851_tech.nc AO          2018-10-11 20:00:26
#>  7 aoml/15852/15852_tech.nc AO          2018-10-11 20:00:28
#>  8 aoml/15853/15853_tech.nc AO          2018-10-11 20:00:29
#>  9 aoml/15854/15854_tech.nc AO          2018-10-11 20:00:30
#> 10 aoml/15855/15855_tech.nc AO          2018-10-11 20:00:34
#> # … with 15,600 more rows

with_argo_example_cache({
  argo_global_traj()
})
#> # A tibble: 17,204 × 8
#>    file      latitude_max latitude_min longitude_max longitude_min profiler_type
#>    <chr>            <dbl>        <dbl>         <dbl>         <dbl>         <dbl>
#>  1 aoml/138…         6.93        0.008        -15.0         -33.8            845
#>  2 aoml/138…         5.21       -0.363         -9.50        -17.8            845
#>  3 aoml/138…         5.93       -0.939        -18.6         -33.7            845
#>  4 aoml/158…        -2.66       -9.19         -16.4         -40.0            845
#>  5 aoml/158…        -1.98       -7.14          -9.90        -35.8            845
#>  6 aoml/158…        -2.73       -6.22           3.33        -21.1            845
#>  7 aoml/158…        -5.04       -8.44          -1.49        -18.9            845
#>  8 aoml/158…        -4.70       -8.25          -4.40        -18.0            845
#>  9 aoml/158…        -4.81       -7.20          -6.41        -12.7            845
#> 10 aoml/158…        -1.79       -6.00           7            -5.45           845
#> # … with 17,194 more rows, and 2 more variables: institution <chr>,
#> #   date_update <dttm>

with_argo_example_cache({
  argo_global_bio_traj()
})
#> Loading argo_global_bio_traj()
#> # A tibble: 557 × 9
#>    file      latitude_max latitude_min longitude_max longitude_min profiler_type
#>    <chr>            <dbl>        <dbl>         <dbl>         <dbl>         <dbl>
#>  1 bodc/390…           NA           NA            NA            NA           836
#>  2 bodc/390…           NA           NA            NA            NA           836
#>  3 bodc/390…           NA           NA            NA            NA           836
#>  4 bodc/390…           NA           NA            NA            NA           836
#>  5 bodc/390…           NA           NA            NA            NA           836
#>  6 bodc/690…           NA           NA            NA            NA           836
#>  7 bodc/690…           NA           NA            NA            NA           836
#>  8 bodc/690…           NA           NA            NA            NA           836
#>  9 bodc/690…           NA           NA            NA            NA           836
#> 10 bodc/690…           NA           NA            NA            NA           836
#> # … with 547 more rows, and 3 more variables: institution <chr>,
#> #   parameters <chr>, date_update <dttm>

if (FALSE) {
# extended download time if no cached version is available
argo_global_prof()
argo_global_bio_prof()
argo_global_synthetic_prof()
}