Load Argo float meta

Load Argo float meta

argo_meta_config_param(path, download = NULL, quiet = FALSE)

argo_meta_missions(path, download = NULL, quiet = FALSE)

argo_meta_trans_system(path, download = NULL, quiet = FALSE)

argo_meta_launch_config_param(path, download = NULL, quiet = FALSE)

argo_meta_sensor(path, download = NULL, quiet = FALSE)

argo_meta_param(path, download = NULL, quiet = FALSE)

argo_read_meta_config_param(file, vars = NULL)

argo_read_meta_missions(file, vars = NULL)

argo_read_meta_trans_system(file, vars = NULL)

argo_read_meta_launch_config_param(file, vars = NULL)

argo_read_meta_sensor(file, vars = NULL)

argo_read_meta_param(file, vars = NULL)

Arguments

path
A path relative to the root directory of argo_mirror().

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

file
A previously downloaded Argo NetCDF file.

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

Value

A tibble::tibble().

Examples

with_argo_example_cache({
  argo_meta_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> # A tibble: 7 x 4
#>   float      config_mission_numb… config_parameter_name     config_parameter_va…
#>   <chr>                     <int> <chr>                                    <dbl>
#> 1 csio/2900…                    1 CONFIG_ParkPressure_dbar                  1500
#> 2 csio/2900…                    1 CONFIG_ProfilePressure_d…                 2000
#> 3 csio/2900…                    1 CONFIG_UpTime_hours                         NA
#> 4 csio/2900…                    1 CONFIG_DownTime_hours                       NA
#> 5 csio/2900…                    1 CONFIG_CycleTime_hours                     240
#> 6 csio/2900…                    1 CONFIG_Direction_NUMBER                      1
#> 7 csio/2900…                    1 CONFIG_MeasureBattery_LO…                    1

with_argo_example_cache({
  argo_meta_missions("dac/csio/2900313/2900313_meta.nc")
})

#> # A tibble: 1 x 3
#>   float        config_mission_number config_mission_comment
#>   <chr>                        <int> <chr>                 
#> 1 csio/2900313                     1 ""                    

with_argo_example_cache({
  argo_meta_trans_system("dac/csio/2900313/2900313_meta.nc")
})

#> # A tibble: 1 x 4
#>   float        trans_system trans_system_id trans_frequency
#>   <chr>        <chr>        <chr>           <chr>          
#> 1 csio/2900313 ARGOS        02528           1/45           

with_argo_example_cache({
  argo_meta_launch_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> # A tibble: 7 x 3
#>   float        launch_config_parameter_name  launch_config_parameter_value
#>   <chr>        <chr>                                                 <dbl>
#> 1 csio/2900313 CONFIG_ParkPressure_dbar                               1500
#> 2 csio/2900313 CONFIG_ProfilePressure_dbar                            2000
#> 3 csio/2900313 CONFIG_UpTime_hours                                      NA
#> 4 csio/2900313 CONFIG_DownTime_hours                                    NA
#> 5 csio/2900313 CONFIG_CycleTime_hours                                  240
#> 6 csio/2900313 CONFIG_Direction_NUMBER                                   1
#> 7 csio/2900313 CONFIG_MeasureBattery_LOGICAL                             1

with_argo_example_cache({
  argo_meta_sensor("dac/csio/2900313/2900313_meta.nc")
})

#> # A tibble: 3 x 5
#>   float        sensor   sensor_maker sensor_model sensor_serial_no
#>   <chr>        <chr>    <chr>        <chr>        <chr>           
#> 1 csio/2900313 CTD_CNDC SBE          SBE41        0327            
#> 2 csio/2900313 CTD_TEMP SBE          SBE41        0327            
#> 3 csio/2900313 CTD_PRES SBE          SBE41        0327            

with_argo_example_cache({
  argo_meta_param("dac/csio/2900313/2900313_meta.nc")
})

#> # A tibble: 3 x 9
#>   float parameter parameter_sensor parameter_units parameter_accur…
#>   <chr> <chr>     <chr>            <chr>           <chr>           
#> 1 csio… PSAL      CTD_CNDC         psu             0.005 psu       
#> 2 csio… TEMP      CTD_TEMP         degC (ITS-90)   0.002 degC      
#> 3 csio… PRES      CTD_PRES         dbar            1.5%            
#> # … with 4 more variables: parameter_resolution <chr>,
#> #   predeployment_calib_equation <chr>, predeployment_calib_coefficient <chr>,
#> #   predeployment_calib_comment <chr>

meta_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_meta.nc",
  package = "argodata"
)

argo_read_meta_config_param(meta_file)

#> # A tibble: 7 x 4
#>   float   CONFIG_MISSION_NU… CONFIG_PARAMETER_NAME             CONFIG_PARAMETER…
#>   <chr>                <int> <chr>                                         <dbl>
#> 1 csio/2…                  1 "CONFIG_ParkPressure_dbar       …              1500
#> 2 csio/2…                  1 "CONFIG_ProfilePressure_dbar    …              2000
#> 3 csio/2…                  1 "CONFIG_UpTime_hours            …                NA
#> 4 csio/2…                  1 "CONFIG_DownTime_hours          …                NA
#> 5 csio/2…                  1 "CONFIG_CycleTime_hours         …               240
#> 6 csio/2…                  1 "CONFIG_Direction_NUMBER        …                 1
#> 7 csio/2…                  1 "CONFIG_MeasureBattery_LOGICAL  …                 1

argo_read_meta_missions(meta_file)

#> # A tibble: 1 x 3
#>   float     CONFIG_MISSION_NUM… CONFIG_MISSION_COMMENT                          
#>   <chr>                   <int> <chr>                                           
#> 1 csio/290…                   1 "                                              …

argo_read_meta_trans_system(meta_file)

#> # A tibble: 1 x 4
#>   float       TRANS_SYSTEM      TRANS_SYSTEM_ID                TRANS_FREQUENCY  
#>   <chr>       <chr>             <chr>                          <chr>            
#> 1 csio/29003… "ARGOS          … "02528                       … "1/45           …

argo_read_meta_launch_config_param(meta_file)

#> # A tibble: 7 x 3
#>   float     LAUNCH_CONFIG_PARAMETER_NAME                  LAUNCH_CONFIG_PARAMET…
#>   <chr>     <chr>                                                          <dbl>
#> 1 csio/290… "CONFIG_ParkPressure_dbar                   …                   1500
#> 2 csio/290… "CONFIG_ProfilePressure_dbar                …                   2000
#> 3 csio/290… "CONFIG_UpTime_hours                        …                     NA
#> 4 csio/290… "CONFIG_DownTime_hours                      …                     NA
#> 5 csio/290… "CONFIG_CycleTime_hours                     …                    240
#> 6 csio/290… "CONFIG_Direction_NUMBER                    …                      1
#> 7 csio/290… "CONFIG_MeasureBattery_LOGICAL              …                      1

argo_read_meta_sensor(meta_file)

#> # A tibble: 3 x 5
#>   float  SENSOR     SENSOR_MAKER          SENSOR_MODEL          SENSOR_SERIAL_NO
#>   <chr>  <chr>      <chr>                 <chr>                 <chr>           
#> 1 csio/… "CTD_CNDC… "SBE                … "SBE41              … "0327          …
#> 2 csio/… "CTD_TEMP… "SBE                … "SBE41              … "0327          …
#> 3 csio/… "CTD_PRES… "SBE                … "SBE41              … "0327          …

argo_read_meta_param(meta_file)

#> # A tibble: 3 x 9
#>   float PARAMETER PARAMETER_SENSOR PARAMETER_UNITS PARAMETER_ACCUR…
#>   <chr> <chr>     <chr>            <chr>           <chr>           
#> 1 csio… "PSAL   … "CTD_CNDC      … "psu          … "0.005 psu     …
#> 2 csio… "TEMP   … "CTD_TEMP      … "degC (ITS-90)… "0.002 degC    …
#> 3 csio… "PRES   … "CTD_PRES      … "dbar         … "1.5%          …
#> # … with 4 more variables: PARAMETER_RESOLUTION <chr>,
#> #   PREDEPLOYMENT_CALIB_EQUATION <chr>, PREDEPLOYMENT_CALIB_COEFFICIENT <chr>,
#> #   PREDEPLOYMENT_CALIB_COMMENT <chr>