Read NetCDF variable metadata from a 'ncdf4' object

Read NetCDF variable metadata from a 'ncdf4' object

argo_nc_read_vars(nc, vars = NULL)

Arguments

nc
A handle created using ncdf4::nc_open().

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

Value

A tibble::tibble() with one row per variable and columns name, longname, units, prec, ndims, size, and dim. Columns size and dim can be unnested (e.g., using tidyr::unnest()) to produce a tibble with one row per variable per dimension.

Examples

prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

nc_prof <- ncdf4::nc_open(prof_file)
argo_nc_read_vars(nc_prof)

#> # A tibble: 64 x 18
#>    name   longname  units prec  ndims size   dim   att_long_name att_conventions
#>    <chr>  <chr>     <chr> <chr> <int> <list> <lis> <chr>         <chr>          
#>  1 DATA_… Data type ""    char      1 <int … <chr… Data type     Argo reference…
#>  2 FORMA… File for… ""    char      1 <int … <chr… File format … NA             
#>  3 HANDB… Data han… ""    char      1 <int … <chr… Data handboo… NA             
#>  4 REFER… Date of … ""    char      1 <int … <chr… Date of refe… YYYYMMDDHHMISS 
#>  5 DATE_… Date of … ""    char      1 <int … <chr… Date of file… YYYYMMDDHHMISS 
#>  6 DATE_… Date of … ""    char      1 <int … <chr… Date of upda… YYYYMMDDHHMISS 
#>  7 PLATF… Float un… ""    char      2 <int … <chr… Float unique… WMO float iden…
#>  8 PLATF… Type of … ""    char      2 <int … <chr… Type of float Argo reference…
#>  9 PROJE… Name of … ""    char      2 <int … <chr… Name of the … NA             
#> 10 PI_NA… Name of … ""    char      2 <int … <chr… Name of the … NA             
#> # … with 54 more rows, and 9 more variables: att_standard_name <chr>,
#> #   att_units <chr>, att_resolution <dbl>, att_axis <chr>, att_valid_min <dbl>,
#> #   att_valid_max <dbl>, att_C_format <chr>, att_FORTRAN_format <chr>,
#> #   att_comment <chr>

ncdf4::nc_close(nc_prof)