Read Argo trajectory NetCDF files

Read Argo trajectory NetCDF files

argo_nc_traj_read(nc, vars = NULL, fill = list())

argo_nc_traj_list_vars(nc)

Arguments

nc
A handle created using ncdf4::nc_open().

vars
A character vector of variable names. These values can be found for each file using argo_nc_traj_list_vars(). If a variable does not exist in nc, a column of NAs is filled in its place.

fill
A list() of missing values that should be applied when vars contains values that are not in argo_nc_prof_list_vars().

Value

A tibble::tibble() containing columns "float" and all columns listed in vars.

Examples

nc_traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)
nc_traj <- ncdf4::nc_open(nc_traj_file)

argo_nc_traj_list_vars(nc_traj)

#>  [1] "JULD"                     "JULD_STATUS"             
#>  [3] "JULD_QC"                  "JULD_ADJUSTED"           
#>  [5] "JULD_ADJUSTED_STATUS"     "JULD_ADJUSTED_QC"        
#>  [7] "LATITUDE"                 "LONGITUDE"               
#>  [9] "POSITION_ACCURACY"        "POSITION_QC"             
#> [11] "CYCLE_NUMBER"             "CYCLE_NUMBER_ADJUSTED"   
#> [13] "MEASUREMENT_CODE"         "PRES"                    
#> [15] "PRES_QC"                  "PRES_ADJUSTED"           
#> [17] "PRES_ADJUSTED_QC"         "PRES_ADJUSTED_ERROR"     
#> [19] "TEMP"                     "TEMP_QC"                 
#> [21] "TEMP_ADJUSTED"            "TEMP_ADJUSTED_QC"        
#> [23] "TEMP_ADJUSTED_ERROR"      "PSAL"                    
#> [25] "PSAL_QC"                  "PSAL_ADJUSTED"           
#> [27] "PSAL_ADJUSTED_QC"         "PSAL_ADJUSTED_ERROR"     
#> [29] "AXES_ERROR_ELLIPSE_MAJOR" "AXES_ERROR_ELLIPSE_MINOR"
#> [31] "AXES_ERROR_ELLIPSE_ANGLE" "SATELLITE_NAME"          

argo_nc_traj_read(nc_traj)

#> # A tibble: 61 x 33
#>    float   JULD JULD_STATUS JULD_QC JULD_ADJUSTED JULD_ADJUSTED_S…
#>    <chr>  <dbl> <chr>       <chr>           <dbl> <chr>           
#>  1 csio… 19573. 0           0                  NA " "             
#>  2 csio… 19573. 4           1                  NA " "             
#>  3 csio… 19582. 4           1                  NA " "             
#>  4 csio… 19583. 4           1                  NA " "             
#>  5 csio… 19583. 4           1                  NA " "             
#>  6 csio… 19583. 4           1                  NA " "             
#>  7 csio… 19583. 4           1                  NA " "             
#>  8 csio… 19583. 4           1                  NA " "             
#>  9 csio… 19592. 4           1                  NA " "             
#> 10 csio… 19592. 4           1                  NA " "             
#> # … with 51 more rows, and 27 more variables: JULD_ADJUSTED_QC <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, POSITION_ACCURACY <chr>,
#> #   POSITION_QC <chr>, CYCLE_NUMBER <int>, CYCLE_NUMBER_ADJUSTED <int>,
#> #   MEASUREMENT_CODE <int>, PRES <dbl>, PRES_QC <chr>, PRES_ADJUSTED <dbl>,
#> #   PRES_ADJUSTED_QC <chr>, PRES_ADJUSTED_ERROR <dbl>, TEMP <dbl>,
#> #   TEMP_QC <chr>, TEMP_ADJUSTED <dbl>, TEMP_ADJUSTED_QC <chr>,
#> #   TEMP_ADJUSTED_ERROR <dbl>, PSAL <dbl>, PSAL_QC <chr>, PSAL_ADJUSTED <dbl>,
#> #   PSAL_ADJUSTED_QC <chr>, PSAL_ADJUSTED_ERROR <dbl>,
#> #   AXES_ERROR_ELLIPSE_MAJOR <dbl>, AXES_ERROR_ELLIPSE_MINOR <dbl>,
#> #   AXES_ERROR_ELLIPSE_ANGLE <dbl>, SATELLITE_NAME <chr>