Transform Argo data to variable-long format

Whereas the variable-wide format returned by most Argo read and load functions are useful, some processing and plotting algorithms need data in a variable-long form, where the variable name is assigned to a column whose value is used to identify the measurement represented by each row.

argo_pivot_longer(tbl, id_cols)

Arguments

tbl
An Argo table with paired value/qc/adjusted columns, likely from argo_prof_levels() or argo_traj_measurement().

id_cols
A vector of column names used to identify each row in the output using dplyr::select() syntax.

Value

A tibble::tibble() with columns id_cols, variable, value, value_qc, value_adjusted, value_adjusted_qc, and value_adjusted_error.

Examples

library(dplyr, warn.conflicts = FALSE)

with_argo_example_cache({
  argo_prof_levels("dac/csio/2902746/profiles/BR2902746_001.nc") %>%
    argo_pivot_longer(id_cols = c(file, date, pres))
})

#> Extracting from 1 file

#> # A tibble: 404 x 9
#>    file  date                 pres variable value value_qc value_adjusted
#>    <chr> <dttm>              <dbl> <chr>    <dbl> <chr>             <dbl>
#>  1 csio… 2018-05-24 05:44:42 0.200 doxy      188. 3                  197.
#>  2 csio… 2018-05-24 05:44:42 0.900 doxy      188. 3                  196.
#>  3 csio… 2018-05-24 05:44:42 1.90  doxy      188. 3                  197.
#>  4 csio… 2018-05-24 05:44:42 2.90  doxy      190. 3                  199.
#>  5 csio… 2018-05-24 05:44:42 4     doxy      190. 3                  199.
#>  6 csio… 2018-05-24 05:44:42 5     doxy      190. 3                  199.
#>  7 csio… 2018-05-24 05:44:42 6     doxy      190. 3                  199.
#>  8 csio… 2018-05-24 05:44:42 7.10  doxy      191. 3                  200.
#>  9 csio… 2018-05-24 05:44:42 8.20  doxy      191. 3                  200.
#> 10 csio… 2018-05-24 05:44:42 9.30  doxy      191. 3                  200.
#> # … with 394 more rows, and 2 more variables: value_adjusted_qc <chr>,
#> #   value_adjusted_error <dbl>