Load Argo profiles

Load Argo profiles

argo_prof_levels(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_prof(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_calib(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_param(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_history(path, vars = NULL, download = NULL, quiet = NA)

Arguments

path
A path relative to the root directory of argo_mirror() or argo_cache_dir(). This value can also be a data.frame with a file column (e.g., a global index as returned by argo_global_meta() and others).

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

Value

A tibble::tibble().

Examples

with_argo_example_cache({
  argo_prof_levels("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> Extracting from 1 file

#> # A tibble: 70 x 18
#>    file            n_levels n_prof   pres pres_qc pres_adjusted pres_adjusted_qc
#>    <chr>              <int>  <int>  <dbl> <chr>           <dbl> <chr>           
#>  1 csio/2900313/p…        1      1   9.80 1                9.80 1               
#>  2 csio/2900313/p…        2      1  20.1  1               20.1  1               
#>  3 csio/2900313/p…        3      1  29.9  1               29.9  1               
#>  4 csio/2900313/p…        4      1  39.7  1               39.7  1               
#>  5 csio/2900313/p…        5      1  49.9  1               49.9  1               
#>  6 csio/2900313/p…        6      1  60.3  1               60.3  1               
#>  7 csio/2900313/p…        7      1  69.7  1               69.7  1               
#>  8 csio/2900313/p…        8      1  80.3  1               80.3  1               
#>  9 csio/2900313/p…        9      1  90.2  1               90.2  1               
#> 10 csio/2900313/p…       10      1 100.   1              100.   1               
#> # … with 60 more rows, and 11 more variables: pres_adjusted_error <dbl>,
#> #   temp <dbl>, temp_qc <chr>, temp_adjusted <dbl>, temp_adjusted_qc <chr>,
#> #   temp_adjusted_error <dbl>, psal <dbl>, psal_qc <chr>, psal_adjusted <dbl>,
#> #   psal_adjusted_qc <chr>, psal_adjusted_error <dbl>

with_argo_example_cache({
  argo_prof_prof("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 x 27
#>   file    n_prof platform_number platform_type project_name pi_name cycle_number
#>   <chr>    <int> <chr>           <chr>         <chr>        <chr>          <dbl>
#> 1 csio/2…      1 2900313         PROVOR        CHINA ARGO … JIANPI…            0
#> # … with 20 more variables: direction <chr>, data_centre <chr>,
#> #   dc_reference <chr>, data_state_indicator <chr>, data_mode <chr>,
#> #   float_serial_no <chr>, firmware_version <chr>, wmo_inst_type <chr>,
#> #   date <dttm>, date_qc <dttm>, date_location <dttm>, latitude <dbl>,
#> #   longitude <dbl>, position_qc <chr>, positioning_system <chr>,
#> #   profile_pres_qc <chr>, profile_temp_qc <chr>, profile_psal_qc <chr>,
#> #   vertical_sampling_scheme <chr>, config_mission_number <dbl>

with_argo_example_cache({
  argo_prof_calib("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 x 9
#>   file    n_param n_calib n_prof parameter scientific_calib_… scientific_calib_…
#>   <chr>     <int>   <int>  <int> <chr>     <chr>              <chr>             
#> 1 csio/2…       1       1      1 PRES      none               none              
#> 2 csio/2…       2       1      1 TEMP      none               none              
#> 3 csio/2…       3       1      1 PSAL      PSAL_ADJUSTED = P… WJO: r =0.9999(+/…
#> # … with 2 more variables: scientific_calib_comment <chr>,
#> #   scientific_calib_date <chr>

with_argo_example_cache({
  argo_prof_param("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 x 4
#>   file                                  n_param n_prof station_parameters
#>   <chr>                                   <int>  <int> <chr>             
#> 1 csio/2900313/profiles/D2900313_000.nc       1      1 PRES              
#> 2 csio/2900313/profiles/D2900313_000.nc       2      1 TEMP              
#> 3 csio/2900313/profiles/D2900313_000.nc       3      1 PSAL              

with_argo_example_cache({
  argo_prof_history("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 x 15
#>   file           n_prof n_history history_institu… history_step history_software
#>   <chr>           <int>     <int> <chr>            <chr>        <chr>           
#> 1 csio/2900313/…      1         1 HZ               ARGQ         ""              
#> 2 csio/2900313/…      1         2 HZ               ARGQ         ""              
#> 3 csio/2900313/…      1         3 HZ               ARSQ         "WJO"           
#> # … with 9 more variables: history_software_release <chr>,
#> #   history_reference <chr>, history_date <chr>, history_action <chr>,
#> #   history_parameter <chr>, history_start_pres <dbl>, history_stop_pres <dbl>,
#> #   history_previous_value <dbl>, history_qctest <chr>