Argo history action reference

This table contains descriptions for the codes used in the history_action column present in argo_prof_history() and argo_traj_history().

argo_reference_history_action

Format

An object of class tbl_df (inherits from tbl, data.frame) with 14 rows and 2 columns.

References

Argo User's Manual, November 2019. https://doi.org/10.13155/29825

Examples

library(dplyr, warn.conflicts = TRUE)

with_argo_example_cache({
  argo_prof_history(
    "dac/csio/2902746/profiles/BR2902746_001.nc",
    vars = "history_action"
  ) %>%
    left_join(
      argo_reference_history_action,
      by = "history_action"
    )
})

#> Extracting from 1 file

#> # A tibble: 4 x 7
#>   file  n_prof cycle_number date                n_history history_action
#>   <chr>  <int>        <int> <dttm>                  <int> <chr>         
#> 1 csio…      1            1 2018-05-24 05:44:42         1 IP            
#> 2 csio…      1            1 2018-05-24 05:44:42         2 IP            
#> 3 csio…      1            1 2018-05-24 05:44:42         3 QCP$          
#> 4 csio…      1            1 2018-05-24 05:44:42         4 QCP$          
#> # … with 1 more variable: history_action_description <chr>

argo_reference_history_action

#> # A tibble: 14 x 2
#>    history_action history_action_description                                    
#>    <chr>          <chr>                                                         
#>  1 CF             Change a quality flag                                         
#>  2 CR             Create record                                                 
#>  3 CV             Change value                                                  
#>  4 DC             Station was checked by duplicate checking software            
#>  5 ED             Edit a parameter value                                        
#>  6 IP             This history group operates on the complete input record      
#>  7 NG             No good trace                                                 
#>  8 PE             Position error. Profile position has been erroneously encoded…
#>  9 QC             Quality Control                                               
#> 10 QCF$           Tests failed                                                  
#> 11 QCP$           Test performed                                                
#> 12 SV             Set a value                                                   
#> 13 TE             Time error. Profile date/time has been erroneously encoded. C…
#> 14 UP             Station passed through the update program