Argo QC test reference

Contains a reference for values in the history_qctest column of argo_prof_history() and argo_traj_history(). The values in this column are a hexadecimal representation of the sum of the qctest_value column. See argo_unnest_history_qctest() to generate the values needed to join this table.

argo_reference_history_qctest

Format

An object of class tbl_df (inherits from tbl, data.frame) with 22 rows and 4 columns.

References

Argo User's Manual, November 2019. https://doi.org/10.13155/29825

Examples

library(dplyr, warn.conflicts = FALSE)

with_argo_example_cache({
  argo_prof_history(
    "dac/csio/2902746/profiles/BR2902746_001.nc",
    vars = "history_qctest"
  ) %>%
    argo_unnest_history_qctest() %>%
    left_join(argo_reference_history_qctest, by = "history_qctest") %>%
    select(history_qctest_description, everything())
})

#> Extracting from 1 file

#> # A tibble: 20 x 9
#>    history_qctest_… file  n_prof cycle_number date                n_history
#>    <chr>            <chr>  <int>        <int> <dttm>                  <int>
#>  1 Platform Identi… csio…      1            1 2018-05-24 05:44:42         1
#>  2 Impossible Date… csio…      1            1 2018-05-24 05:44:42         1
#>  3 Impossible Loca… csio…      1            1 2018-05-24 05:44:42         1
#>  4 Position on Lan… csio…      1            1 2018-05-24 05:44:42         1
#>  5 Impossible Spee… csio…      1            1 2018-05-24 05:44:42         1
#>  6 Global Range te… csio…      1            1 2018-05-24 05:44:42         1
#>  7 Regional Global… csio…      1            1 2018-05-24 05:44:42         1
#>  8 Pressure Increa… csio…      1            1 2018-05-24 05:44:42         1
#>  9 Spike test       csio…      1            1 2018-05-24 05:44:42         1
#> 10 Gradient test    csio…      1            1 2018-05-24 05:44:42         1
#> 11 Digit Rollover … csio…      1            1 2018-05-24 05:44:42         1
#> 12 Stuck Value test csio…      1            1 2018-05-24 05:44:42         1
#> 13 Density Inversi… csio…      1            1 2018-05-24 05:44:42         1
#> 14 Gross Salinity … csio…      1            1 2018-05-24 05:44:42         1
#> 15 Frozen profile … csio…      1            1 2018-05-24 05:44:42         1
#> 16 Deepest pressur… csio…      1            1 2018-05-24 05:44:42         1
#> 17 Global Range te… csio…      1            1 2018-05-24 05:44:42         3
#> 18 Spike test       csio…      1            1 2018-05-24 05:44:42         3
#> 19 Gradient test    csio…      1            1 2018-05-24 05:44:42         3
#> 20 Frozen profile … csio…      1            1 2018-05-24 05:44:42         3
#> # … with 3 more variables: history_qctest <chr>, history_qctest_number <int>,
#> #   history_qctest_flag <int>

argo_reference_history_qctest

#> # A tibble: 22 x 4
#>    history_qctest_num… history_qctest_f… history_qctest_descript… history_qctest
#>                  <int>             <int> <chr>                    <chr>         
#>  1                   1                 2 Platform Identification… 2             
#>  2                   2                 4 Impossible Date test     4             
#>  3                   3                 8 Impossible Location test 8             
#>  4                   4                16 Position on Land test    10            
#>  5                   5                32 Impossible Speed test    20            
#>  6                   6                64 Global Range test        40            
#>  7                   7               128 Regional Global Paramet… 80            
#>  8                   8               256 Pressure Increasing test 100           
#>  9                   9               512 Spike test               200           
#> 10                  10              1024 Top and Bottom Spike te… 400           
#> # … with 12 more rows