Argo QC flag reference

This table contains the quality control descriptions of the one-character flags that appear in variables that end with _qc. These variables appear in argo_prof_levels(), argo_prof_prof(), and argo_traj_measurement(). See also argo_reference_history_qctest for information about specific tests.

argo_reference_qc_flag

Format

An object of class tbl_df (inherits from tbl, data.frame) with 10 rows and 4 columns.

References

Argo User's Manual, November 2019. doi: 10.13155/29825

Examples

library(dplyr, warn.conflicts = FALSE)

with_argo_example_cache({
  argo_prof_levels(
    "dac/csio/2902746/profiles/BR2902746_001.nc",
    vars = c("pres", "pres_qc", "doxy", "doxy_qc")
  ) %>%
    dplyr::left_join(
      argo_reference_qc_flag,
      by = c("doxy_qc" = "qc_flag")
    )
})
#> Extracting from 1 file
#> Warning: '/Users/runner/work/_temp/Library/argodata/cache-test/dac/csio/2902746/profiles/BR2902746_001.nc' is missing variable:
#> 'PRES_QC'
#> # A tibble: 101 × 9
#>    file    n_levels n_prof  pres  doxy doxy_qc qc_description qc_comment_realti…
#>    <chr>      <int>  <int> <dbl> <dbl> <chr>   <chr>          <chr>             
#>  1 csio/2…        1      1 0.200  188. 3       Bad data that… Test 15 or Test 1…
#>  2 csio/2…        2      1 0.900  188. 3       Bad data that… Test 15 or Test 1…
#>  3 csio/2…        3      1 1.90   188. 3       Bad data that… Test 15 or Test 1…
#>  4 csio/2…        4      1 2.90   190. 3       Bad data that… Test 15 or Test 1…
#>  5 csio/2…        5      1 4      190. 3       Bad data that… Test 15 or Test 1…
#>  6 csio/2…        6      1 5      190. 3       Bad data that… Test 15 or Test 1…
#>  7 csio/2…        7      1 6      190. 3       Bad data that… Test 15 or Test 1…
#>  8 csio/2…        8      1 7.10   191. 3       Bad data that… Test 15 or Test 1…
#>  9 csio/2…        9      1 8.20   191. 3       Bad data that… Test 15 or Test 1…
#> 10 csio/2…       10      1 9.30   191. 3       Bad data that… Test 15 or Test 1…
#> # … with 91 more rows, and 1 more variable: qc_comment_delayed <chr>

argo_reference_qc_flag
#> # A tibble: 10 × 4
#>    qc_flag qc_description     qc_comment_realtime       qc_comment_delayed      
#>    <chr>   <chr>              <chr>                     <chr>                   
#>  1 0       No QC performed    No QC performed           No QC performed         
#>  2 1       Good data          All real time QC tests p… The adjusted value is s…
#>  3 2       Probably good data NA                        Probably good           
#>  4 3       Bad data that are… Test 15 or Test 16 or Te… An adjustment has been …
#>  5 4       Bad data           Data have failed one or … Bad data.  Not adjustab…
#>  6 5       Value changed      Value changed             Value changed           
#>  7 6       Not currently used Not currently used        Not currently used      
#>  8 7       Not currently used Not currently used        Not currently used      
#>  9 8       Estimated          Estimated value (interpo… Estimated value (interp…
#> 10 9       Missing value      Missing value             Missing value