Load Argo float meta

Use argo_tech_tech_param() to extract information from Argo technical spec NetCDF files. Use argo_read_tech_tech_param() to extract information from a single previously-downloaded NetCDF file.

argo_tech_tech_param(path, download = NULL, quiet = NA)

Arguments

path

A path relative to the root directory of argo_mirror() or argo_cache_dir(). This value can also be a data.frame with a file column (e.g., a global index as returned by argo_global_meta() and others).

download

A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet

Use FALSE to show which files are downloaded and for more verbose error messages.

Value

A tibble::tibble() with one row per technical parameter per cycle and columns file, n_tech_param, technical_parameter_name, technical_parameter_value, and cycle_number.

Examples

with_argo_example_cache({
  argo_tech_tech_param("dac/csio/2900313/2900313_tech.nc")
})
#> Extracting from 1 file
#> # A tibble: 96 × 5
#>    file      n_tech_param technical_parameter_na… technical_parame… cycle_number
#>    <chr>            <int> <chr>                   <chr>                    <dbl>
#>  1 csio/290…            1 NUMBER_ValveActionsAtS… 150                          0
#>  2 csio/290…            2 CLOCK_InitialStabiliza… 0.2                          0
#>  3 csio/290…            3 NUMBER_PumpActionsDuri… 0                            0
#>  4 csio/290…            4 NUMBER_RepositionsDuri… 0                            0
#>  5 csio/290…            5 NUMBER_PumpActionsDuri… 14                           0
#>  6 csio/290…            6 NUMBER_PumpActionsAtSu… 2                            0
#>  7 csio/290…            7 PRES_SurfaceOffsetCorr… 1.025                        0
#>  8 csio/290…            8 PRESSURE_InternalVacuu… 740.0                        0
#>  9 csio/290…            9 NUMBER_AscentArgosMess… 14                           0
#> 10 csio/290…           10 NUMBER_AscentSamples_C… 71                           0
#> # … with 86 more rows