Read Argo NetCDF variable metadata

Read Argo NetCDF variable metadata

argo_vars(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_vars(file, vars = NULL)

argo_nc_read_vars(nc, vars = NULL)

Arguments

path
A path relative to the root directory of argo_mirror().

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache or global files that were cached more than 24 hours ago.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

file
A previously downloaded Argo NetCDF file.

nc
A handle created using ncdf4::nc_open().

Value

A tibble::tibble() with one row per variable and columns name, longname, units, prec, ndims, size, and dim. Columns size and dim can be unnested (e.g., using tidyr::unnest()) to produce a tibble with one row per variable per dimension.

Examples

with_argo_example_cache({
  argo_vars("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> # A tibble: 64 x 8
#>    float     name         longname              units prec  ndims size    dim   
#>    <chr>     <chr>        <chr>                 <chr> <chr> <int> <list>  <list>
#>  1 csio/290… data_type    Data type             ""    char      1 <int [… <chr …
#>  2 csio/290… format_vers… File format version   ""    char      1 <int [… <chr …
#>  3 csio/290… handbook_ve… Data handbook version ""    char      1 <int [… <chr …
#>  4 csio/290… reference_d… Date of reference fo… ""    char      1 <int [… <chr …
#>  5 csio/290… date_creati… Date of file creation ""    char      1 <int [… <chr …
#>  6 csio/290… date_update  Date of update of th… ""    char      1 <int [… <chr …
#>  7 csio/290… platform_nu… Float unique identif… ""    char      2 <int [… <chr …
#>  8 csio/290… platform_ty… Type of float         ""    char      2 <int [… <chr …
#>  9 csio/290… project_name Name of the project   ""    char      2 <int [… <chr …
#> 10 csio/290… pi_name      Name of the principa… ""    char      2 <int [… <chr …
#> # … with 54 more rows

prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_vars(prof_file)

#> # A tibble: 64 x 8
#>    float     name         longname              units prec  ndims size    dim   
#>    <chr>     <chr>        <chr>                 <chr> <chr> <int> <list>  <list>
#>  1 csio/290… DATA_TYPE    Data type             ""    char      1 <int [… <chr …
#>  2 csio/290… FORMAT_VERS… File format version   ""    char      1 <int [… <chr …
#>  3 csio/290… HANDBOOK_VE… Data handbook version ""    char      1 <int [… <chr …
#>  4 csio/290… REFERENCE_D… Date of reference fo… ""    char      1 <int [… <chr …
#>  5 csio/290… DATE_CREATI… Date of file creation ""    char      1 <int [… <chr …
#>  6 csio/290… DATE_UPDATE  Date of update of th… ""    char      1 <int [… <chr …
#>  7 csio/290… PLATFORM_NU… Float unique identif… ""    char      2 <int [… <chr …
#>  8 csio/290… PLATFORM_TY… Type of float         ""    char      2 <int [… <chr …
#>  9 csio/290… PROJECT_NAME Name of the project   ""    char      2 <int [… <chr …
#> 10 csio/290… PI_NAME      Name of the principa… ""    char      2 <int [… <chr …
#> # … with 54 more rows

nc_prof <- ncdf4::nc_open(prof_file)
argo_nc_read_vars(nc_prof)

#> # A tibble: 64 x 8
#>    float     name         longname              units prec  ndims size    dim   
#>    <chr>     <chr>        <chr>                 <chr> <chr> <int> <list>  <list>
#>  1 csio/290… DATA_TYPE    Data type             ""    char      1 <int [… <chr …
#>  2 csio/290… FORMAT_VERS… File format version   ""    char      1 <int [… <chr …
#>  3 csio/290… HANDBOOK_VE… Data handbook version ""    char      1 <int [… <chr …
#>  4 csio/290… REFERENCE_D… Date of reference fo… ""    char      1 <int [… <chr …
#>  5 csio/290… DATE_CREATI… Date of file creation ""    char      1 <int [… <chr …
#>  6 csio/290… DATE_UPDATE  Date of update of th… ""    char      1 <int [… <chr …
#>  7 csio/290… PLATFORM_NU… Float unique identif… ""    char      2 <int [… <chr …
#>  8 csio/290… PLATFORM_TY… Type of float         ""    char      2 <int [… <chr …
#>  9 csio/290… PROJECT_NAME Name of the project   ""    char      2 <int [… <chr …
#> 10 csio/290… PI_NAME      Name of the principa… ""    char      2 <int [… <chr …
#> # … with 54 more rows

ncdf4::nc_close(nc_prof)