Select rows of Argo tables

These functions are intended to be applied to argo_global_meta() and other global index tables in the argo_global_*() family.

argo_filter_radius(tbl, latitude, longitude, radius_km)

argo_filter_rect(tbl, latitude_min, latitude_max, longitude_min, longitude_max)

argo_filter_date(tbl, date_min, date_max = Sys.time())

argo_filter_updated(tbl, date_update_min, date_update_max = Sys.time())

argo_filter_float(tbl, float)

argo_filter_data_mode(tbl, data_mode)

Arguments

tbl
A data.frame, ideally derived from argo_global_meta() and family. The column conventions used by the global indexes is assumed (e.g., columns latitude and longitude exist).

latitude, longitude
A location.

radius_km
A radius from the point described by latitude and longitude.

latitude_max, latitude_min, longitude_max, longitude_min
A rectangle describing the desired bounds. A rectangle where longitude_min is greater than longitude_max are interpreted as wrapping across the international date line.

date_min, date_max, date_update_min, date_update_max
A range of datetimes. Users are responsible for setting the timezone for these objects and are encouraged to used UTC.

float
A float identifier.

data_mode
One of "realtime" or "delayed".

Value

tbl with rows that match the search criteria.