Select rows of Argo tables

These functions are intended to be applied to argo_global_meta() and other global index tables in the argo_global_*() family.

argo_filter_radius(tbl, latitude, longitude, radius_km)

argo_filter_rect(tbl, latitude_min, latitude_max, longitude_min, longitude_max)

argo_filter_date(tbl, date_min, date_max = Sys.time())

argo_filter_updated(tbl, date_update_min, date_update_max = Sys.time())

argo_filter_float(tbl, float)

argo_filter_data_mode(tbl, data_mode)

Arguments

tbl
A data.frame, ideally derived from argo_global_meta() and family. The column conventions used by the global indexes is assumed (e.g., columns latitude and longitude exist).

latitude, longitude
A location.

radius_km
A radius from the point described by latitude and longitude.

latitude_max, latitude_min, longitude_max, longitude_min
A rectangle describing the desired bounds. A rectangle where longitude_min is greater than longitude_max are interpreted as wrapping across the international date line.

date_min, date_max, date_update_min, date_update_max
A range of datetimes. Users are responsible for setting the timezone for these objects and are encouraged to used UTC.

float
A float identifier.

data_mode
One of "realtime" or "delayed".

Value

tbl with rows that match the search criteria.

Examples

library(dplyr, warn.conflicts = FALSE)

# \dontrun{
argo_global_prof() %>%
  # within 500 km of Halifax, Nova Scotia
  argo_filter_radius(45, -64, 500)

#> Loading argo_global_prof()

#> Downloading 1 file from 'https://data-argo.ifremer.fr'

#> # A tibble: 2,267 x 8
#>    file  date                latitude longitude ocean profiler_type institution
#>    <chr> <dttm>                 <dbl>     <dbl> <chr>         <dbl> <chr>      
#>  1 aoml… 2013-04-20 10:47:20     43.0     -59.3 A               851 AO         
#>  2 aoml… 2013-04-30 11:27:49     43.0     -58.7 A               851 AO         
#>  3 aoml… 2013-05-10 11:00:04     42.9     -59.5 A               851 AO         
#>  4 aoml… 2013-05-20 10:14:33     43.0     -59.1 A               851 AO         
#>  5 aoml… 2013-05-30 10:34:48     42.8     -58.8 A               851 AO         
#>  6 aoml… 2003-07-23 08:26:00     40.6     -64.3 A               852 AO         
#>  7 aoml… 2005-04-03 07:40:04     41.4     -67.5 A               852 AO         
#>  8 aoml… 2005-04-13 07:38:45     41.6     -67.5 A               852 AO         
#>  9 aoml… 2005-04-23 07:35:52     41.9     -67.6 A               852 AO         
#> 10 aoml… 2004-06-27 05:53:00     42.1     -59.4 A               852 AO         
#> # … with 2,257 more rows, and 1 more variable: date_update <dttm>

# }

with_argo_example_cache({
  argo_global_traj() %>%
    argo_filter_rect(40, 60, -64, -54)
})

#> Loading argo_global_traj()

#> # A tibble: 669 x 8
#>    file  latitude_max latitude_min longitude_max longitude_min profiler_type
#>    <chr>        <dbl>        <dbl>         <dbl>         <dbl>         <dbl>
#>  1 aoml…         52.6        33.5          -40.1         -74.0           851
#>  2 aoml…         40.5        25.6          -57.3         -75.3           851
#>  3 aoml…         41.8         1.42         180.         -180.            846
#>  4 aoml…         42.4        28.4          -47.1         -75.3           851
#>  5 aoml…         40.9        27.7          -49.9         -76.6           852
#>  6 aoml…         45.6        19.9          -27.8         -75.9           851
#>  7 aoml…         41.8        33.8          -44.1         -74.3           854
#>  8 aoml…         58.0        52.9          180.         -180.            846
#>  9 aoml…         60.4        53.9          180.         -180.            846
#> 10 aoml…         54.6        51.1          180.         -180.            846
#> # … with 659 more rows, and 2 more variables: institution <chr>,
#> #   date_update <dttm>

with_argo_example_cache({
  argo_global_traj() %>%
    argo_filter_updated("2020-01-01 00:00") %>%
    select(date_update, everything())
})

#> # A tibble: 5,695 x 8
#>    date_update         file  latitude_max latitude_min longitude_max
#>    <dttm>              <chr>        <dbl>        <dbl>         <dbl>
#>  1 2020-06-25 20:34:53 aoml…        21.0         17.1           40.7
#>  2 2020-06-25 20:35:16 aoml…        21.7         19.5           38.7
#>  3 2020-08-04 20:00:59 aoml…       -42.3        -61.3          177. 
#>  4 2020-04-10 20:01:01 aoml…       -38.9        -48.8          143. 
#>  5 2020-01-20 21:01:32 aoml…       -23.5        -40.8           81.3
#>  6 2020-03-06 21:01:34 aoml…       -21.9        -49.7          143. 
#>  7 2020-01-16 21:01:19 aoml…         2.84        -2.66          98.6
#>  8 2020-10-14 20:02:02 aoml…        -3.62        -9.50          94.7
#>  9 2020-10-13 20:02:00 aoml…        -1.54        -9.69         107. 
#> 10 2020-10-14 20:02:12 aoml…         1.73       -12.9          112. 
#> # … with 5,685 more rows, and 3 more variables: longitude_min <dbl>,
#> #   profiler_type <dbl>, institution <chr>

with_argo_example_cache({
  argo_global_traj() %>%
    argo_filter_float(c("13857", "15851"))
})

#> # A tibble: 2 x 8
#>   file  latitude_max latitude_min longitude_max longitude_min profiler_type
#>   <chr>        <dbl>        <dbl>         <dbl>         <dbl>         <dbl>
#> 1 aoml…         6.93        0.008        -15.0          -33.8           845
#> 2 aoml…        -2.73       -6.22           3.33         -21.1           845
#> # … with 2 more variables: institution <chr>, date_update <dttm>

with_argo_example_cache({
  argo_global_traj() %>%
    argo_filter_data_mode("delayed")
})

#> # A tibble: 1,986 x 8
#>    file  latitude_max latitude_min longitude_max longitude_min profiler_type
#>    <chr>        <dbl>        <dbl>         <dbl>         <dbl>         <dbl>
#>  1 aoml…         6.56       -9.67          -14.0         -50.6           851
#>  2 aoml…        -1.56       -6.07          -18.6         -43.2           851
#>  3 aoml…         5.13        0.824         -27.4         -35.4           851
#>  4 aoml…       -35.6       -47.1            80.4          21.1           851
#>  5 aoml…       -33.5       -40.3            41.1          18.1           851
#>  6 aoml…       -34.7       -53.9            95.6          20.8           851
#>  7 aoml…       -25.7       -44.0            42.6          16.4           851
#>  8 aoml…       -22.3       -28.7            83.3          73.9           851
#>  9 aoml…       -19.8       -41.1            72.0          16.8           851
#> 10 aoml…       -22.8       -32.1            67.0          47.2           851
#> # … with 1,976 more rows, and 2 more variables: institution <chr>,
#> #   date_update <dttm>