Load Argo global index files

Load Argo global index files

argo_global_meta(download = NULL, quiet = FALSE)

argo_global_prof(download = NULL, quiet = FALSE)

argo_global_tech(download = NULL, quiet = FALSE)

argo_global_traj(download = NULL, quiet = FALSE)

argo_global_bio_traj(download = NULL, quiet = FALSE)

argo_global_bio_prof(download = NULL, quiet = FALSE)

argo_global_synthetic_prof(download = NULL, quiet = FALSE)

Arguments

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

Value

A tibble::tibble().

Examples

with_argo_example_cache({
  argo_global_meta()
})

#> Loading argo_global_meta()

#> # A tibble: 16,062 x 4
#>    file                     profiler_type institution date_update        
#>    <chr>                            <dbl> <chr>       <dttm>             
#>  1 aoml/13857/13857_meta.nc           845 AO          2018-10-11 20:00:14
#>  2 aoml/13858/13858_meta.nc           845 AO          2018-10-11 20:00:15
#>  3 aoml/13859/13859_meta.nc           845 AO          2018-10-11 20:00:25
#>  4 aoml/15819/15819_meta.nc           845 AO          2018-10-11 20:00:16
#>  5 aoml/15820/15820_meta.nc           845 AO          2018-10-11 20:00:18
#>  6 aoml/15851/15851_meta.nc           845 AO          2018-10-11 20:00:26
#>  7 aoml/15852/15852_meta.nc           845 AO          2018-10-11 20:00:28
#>  8 aoml/15853/15853_meta.nc           845 AO          2018-10-11 20:00:29
#>  9 aoml/15854/15854_meta.nc           845 AO          2018-10-11 20:00:30
#> 10 aoml/15855/15855_meta.nc           845 AO          2018-10-11 20:00:34
#> # … with 16,052 more rows

with_argo_example_cache({
  argo_global_tech()
})

#> Loading argo_global_tech()

#> # A tibble: 15,610 x 3
#>    file                     institution date_update        
#>    <chr>                    <chr>       <dttm>             
#>  1 aoml/13857/13857_tech.nc AO          2018-10-11 20:00:14
#>  2 aoml/13858/13858_tech.nc AO          2018-10-11 20:00:15
#>  3 aoml/13859/13859_tech.nc AO          2018-10-11 20:00:25
#>  4 aoml/15819/15819_tech.nc AO          2018-10-11 20:00:16
#>  5 aoml/15820/15820_tech.nc AO          2018-10-11 20:00:18
#>  6 aoml/15851/15851_tech.nc AO          2018-10-11 20:00:26
#>  7 aoml/15852/15852_tech.nc AO          2018-10-11 20:00:28
#>  8 aoml/15853/15853_tech.nc AO          2018-10-11 20:00:29
#>  9 aoml/15854/15854_tech.nc AO          2018-10-11 20:00:30
#> 10 aoml/15855/15855_tech.nc AO          2018-10-11 20:00:34
#> # … with 15,600 more rows

with_argo_example_cache({
  argo_global_traj()
})

#> # A tibble: 17,204 x 8
#>    file  latitude_max latitude_min longitude_max longitude_min profiler_type
#>    <chr>        <dbl>        <dbl>         <dbl>         <dbl>         <dbl>
#>  1 aoml…         6.93        0.008        -15.0         -33.8            845
#>  2 aoml…         5.21       -0.363         -9.50        -17.8            845
#>  3 aoml…         5.93       -0.939        -18.6         -33.7            845
#>  4 aoml…        -2.66       -9.19         -16.4         -40.0            845
#>  5 aoml…        -1.98       -7.14          -9.90        -35.8            845
#>  6 aoml…        -2.73       -6.22           3.33        -21.1            845
#>  7 aoml…        -5.04       -8.44          -1.49        -18.9            845
#>  8 aoml…        -4.70       -8.25          -4.40        -18.0            845
#>  9 aoml…        -4.81       -7.20          -6.41        -12.7            845
#> 10 aoml…        -1.79       -6.00           7            -5.45           845
#> # … with 17,194 more rows, and 2 more variables: institution <chr>,
#> #   date_update <dttm>

with_argo_example_cache({
  argo_global_bio_traj()
})

#> Loading argo_global_bio_traj()

#> # A tibble: 557 x 9
#>    file  latitude_max latitude_min longitude_max longitude_min profiler_type
#>    <chr>        <dbl>        <dbl>         <dbl>         <dbl>         <dbl>
#>  1 bodc…           NA           NA            NA            NA           836
#>  2 bodc…           NA           NA            NA            NA           836
#>  3 bodc…           NA           NA            NA            NA           836
#>  4 bodc…           NA           NA            NA            NA           836
#>  5 bodc…           NA           NA            NA            NA           836
#>  6 bodc…           NA           NA            NA            NA           836
#>  7 bodc…           NA           NA            NA            NA           836
#>  8 bodc…           NA           NA            NA            NA           836
#>  9 bodc…           NA           NA            NA            NA           836
#> 10 bodc…           NA           NA            NA            NA           836
#> # … with 547 more rows, and 3 more variables: institution <chr>,
#> #   parameters <chr>, date_update <dttm>

# \dontrun{
# extended download time if no cached version is available
argo_global_prof()

#> # A tibble: 2,376,147 x 8
#>    file  date                latitude longitude ocean profiler_type institution
#>    <chr> <dttm>                 <dbl>     <dbl> <chr>         <dbl> <chr>      
#>  1 aoml… 1997-07-29 20:03:00    0.267     -16.0 A               845 AO         
#>  2 aoml… 1997-08-09 19:21:12    0.072     -17.7 A               845 AO         
#>  3 aoml… 1997-08-20 18:45:45    0.543     -19.6 A               845 AO         
#>  4 aoml… 1997-08-31 19:39:05    1.26      -20.5 A               845 AO         
#>  5 aoml… 1997-09-11 18:58:08    0.72      -20.8 A               845 AO         
#>  6 aoml… 1997-09-22 19:57:02    1.76      -21.6 A               845 AO         
#>  7 aoml… 1997-10-03 19:15:49    2.60      -21.6 A               845 AO         
#>  8 aoml… 1997-10-14 18:39:35    1.76      -21.6 A               845 AO         
#>  9 aoml… 1997-10-25 19:32:34    1.80      -21.8 A               845 AO         
#> 10 aoml… 1997-11-05 18:51:42    1.64      -21.4 A               845 AO         
#> # … with 2,376,137 more rows, and 1 more variable: date_update <dttm>

argo_global_bio_prof()

#> Loading argo_global_bio_prof()

#> Downloading 1 file from 'https://data-argo.ifremer.fr'

#> # A tibble: 226,211 x 10
#>    file  date                latitude longitude ocean profiler_type institution
#>    <chr> <dttm>                 <dbl>     <dbl> <chr>         <dbl> <chr>      
#>  1 aoml… 2006-10-22 02:16:24    -40.3      73.4 I               846 AO         
#>  2 aoml… 2006-11-01 06:44:23    -40.4      73.5 I               846 AO         
#>  3 aoml… 2006-11-11 10:12:22    -40.5      73.3 I               846 AO         
#>  4 aoml… 2006-11-21 07:50:21    -40.1      73.1 I               846 AO         
#>  5 aoml… 2006-12-01 18:33:00    -39.6      73.2 I               846 AO         
#>  6 aoml… 2006-12-11 22:27:04    -39.4      73.5 I               846 AO         
#>  7 aoml… 2006-12-22 05:20:18    -39.8      74.0 I               846 AO         
#>  8 aoml… 2007-01-01 06:34:17    -40.0      74.2 I               846 AO         
#>  9 aoml… 2007-01-11 12:14:56    -39.4      75.3 I               846 AO         
#> 10 aoml… 2007-01-21 04:22:15    -38.5      76.0 I               846 AO         
#> # … with 226,201 more rows, and 3 more variables: parameters <chr>,
#> #   parameter_data_mode <chr>, date_update <dttm>

argo_global_synthetic_prof()

#> Loading argo_global_synthetic_prof()

#> Downloading 1 file from 'https://data-argo.ifremer.fr'

#> # A tibble: 211,345 x 10
#>    file  date                latitude longitude ocean profiler_type institution
#>    <chr> <dttm>                 <dbl>     <dbl> <chr>         <dbl> <chr>      
#>  1 aoml… 2006-10-22 02:16:24    -40.3      73.4 I               846 AO         
#>  2 aoml… 2006-11-01 06:44:23    -40.4      73.5 I               846 AO         
#>  3 aoml… 2006-11-11 10:12:22    -40.5      73.3 I               846 AO         
#>  4 aoml… 2006-11-21 07:50:21    -40.1      73.1 I               846 AO         
#>  5 aoml… 2006-12-01 18:33:00    -39.6      73.2 I               846 AO         
#>  6 aoml… 2006-12-11 22:27:04    -39.4      73.5 I               846 AO         
#>  7 aoml… 2006-12-22 05:20:18    -39.8      74.0 I               846 AO         
#>  8 aoml… 2007-01-01 06:34:17    -40.0      74.2 I               846 AO         
#>  9 aoml… 2007-01-11 12:14:56    -39.4      75.3 I               846 AO         
#> 10 aoml… 2007-01-21 04:22:15    -38.5      76.0 I               846 AO         
#> # … with 211,335 more rows, and 3 more variables: parameters <chr>,
#> #   parameter_data_mode <chr>, date_update <dttm>

# }