Load Argo float meta

Load Argo float meta

argo_meta_config_param(path, download = NULL, quiet = NA)

argo_meta_missions(path, download = NULL, quiet = NA)

argo_meta_trans_system(path, download = NULL, quiet = NA)

argo_meta_positioning_system(path, download = NULL, quiet = NA)

argo_meta_launch_config_param(path, download = NULL, quiet = NA)

argo_meta_sensor(path, download = NULL, quiet = NA)

argo_meta_param(path, download = NULL, quiet = NA)

Arguments

path
A path relative to the root directory of argo_mirror() or argo_cache_dir(). This value can also be a data.frame with a file column (e.g., a global index as returned by argo_global_meta() and others).

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

Value

A tibble::tibble().

Examples

with_argo_example_cache({
  argo_meta_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 7 x 5
#>   file          n_config_param n_missions config_parameter… config_parameter_na…
#>   <chr>                  <int>      <int>             <dbl> <chr>               
#> 1 csio/2900313…              1          1              1500 CONFIG_ParkPressure…
#> 2 csio/2900313…              2          1              2000 CONFIG_ProfilePress…
#> 3 csio/2900313…              3          1                NA CONFIG_UpTime_hours 
#> 4 csio/2900313…              4          1                NA CONFIG_DownTime_hou…
#> 5 csio/2900313…              5          1               240 CONFIG_CycleTime_ho…
#> 6 csio/2900313…              6          1                 1 CONFIG_Direction_NU…
#> 7 csio/2900313…              7          1                 1 CONFIG_MeasureBatte…

with_argo_example_cache({
  argo_meta_missions("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 x 4
#>   file                     n_missions config_mission_number config_mission_comm…
#>   <chr>                         <int>                 <dbl> <chr>               
#> 1 csio/2900313/2900313_me…          1                     1 ""                  

with_argo_example_cache({
  argo_meta_trans_system("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 x 5
#>   file               n_trans_system trans_system trans_system_id trans_frequency
#>   <chr>                       <int> <chr>        <chr>           <chr>          
#> 1 csio/2900313/2900…              1 ARGOS        02528           1/45           

with_argo_example_cache({
  argo_meta_positioning_system("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 x 3
#>   file                         n_positioning_system positioning_system
#>   <chr>                                       <int> <chr>             
#> 1 csio/2900313/2900313_meta.nc                    1 ARGOS             

with_argo_example_cache({
  argo_meta_launch_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 7 x 4
#>   file            n_launch_config_p… launch_config_parame… launch_config_parame…
#>   <chr>                        <int> <chr>                                 <dbl>
#> 1 csio/2900313/2…                  1 CONFIG_ParkPressure_…                  1500
#> 2 csio/2900313/2…                  2 CONFIG_ProfilePressu…                  2000
#> 3 csio/2900313/2…                  3 CONFIG_UpTime_hours                      NA
#> 4 csio/2900313/2…                  4 CONFIG_DownTime_hours                    NA
#> 5 csio/2900313/2…                  5 CONFIG_CycleTime_hou…                   240
#> 6 csio/2900313/2…                  6 CONFIG_Direction_NUM…                     1
#> 7 csio/2900313/2…                  7 CONFIG_MeasureBatter…                     1

with_argo_example_cache({
  argo_meta_sensor("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 x 6
#>   file               n_sensor sensor  sensor_maker sensor_model sensor_serial_no
#>   <chr>                 <int> <chr>   <chr>        <chr>        <chr>           
#> 1 csio/2900313/2900…        1 CTD_CN… SBE          SBE41        0327            
#> 2 csio/2900313/2900…        2 CTD_TE… SBE          SBE41        0327            
#> 3 csio/2900313/2900…        3 CTD_PR… SBE          SBE41        0327            

with_argo_example_cache({
  argo_meta_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 x 10
#>   file       n_param parameter parameter_sensor parameter_units parameter_accur…
#>   <chr>        <int> <chr>     <chr>            <chr>           <chr>           
#> 1 csio/2900…       1 PSAL      CTD_CNDC         psu             0.005 psu       
#> 2 csio/2900…       2 TEMP      CTD_TEMP         degC (ITS-90)   0.002 degC      
#> 3 csio/2900…       3 PRES      CTD_PRES         dbar            1.5%            
#> # … with 4 more variables: parameter_resolution <chr>,
#> #   predeployment_calib_equation <chr>, predeployment_calib_coefficient <chr>,
#> #   predeployment_calib_comment <chr>