Load Argo float meta

Use argo_meta_*() functions to extract information from Argo meta NetCDF files. Use argo_read_meta_*() to extract information from a single previously-downloaded NetCDF file. Using argo_info() on meta files is also useful for extracting general float information.

argo_meta_config_param(path, download = NULL, quiet = NA)

argo_meta_missions(path, download = NULL, quiet = NA)

argo_meta_trans_system(path, download = NULL, quiet = NA)

argo_meta_positioning_system(path, download = NULL, quiet = NA)

argo_meta_launch_config_param(path, download = NULL, quiet = NA)

argo_meta_sensor(path, download = NULL, quiet = NA)

argo_meta_param(path, download = NULL, quiet = NA)

Arguments

path
A path relative to the root directory of argo_mirror() or argo_cache_dir(). This value can also be a data.frame with a file column (e.g., a global index as returned by argo_global_meta() and others).

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

Value

A tibble::tibble() with

argo_meta_config_param(): one row per file per configuration parameter and columns file, n_config_param, n_missions, config_parameter_value, and config_parameter_name.

argo_meta_missions(): one row per file per mission and columns file, n_missions, config_mission_number, and config_mission_comment.

argo_meta_trans_system(): one row per file per transmission system and columns file, n_trans_system, trans_system, trans_system_id, and trans_frequency.

argo_meta_positioning_system(): one row per file per positioning system and columns file, n_positioning_system, and positioning_system.

argo_meta_launch_config_param(): one row per file per launch configuration parameter and columns file, n_launch_config_param, launch_config_parameter_name, and launch_config_parameter_value.

argo_meta_sensor(): one row per file per sensor and columns file, n_sensor, sensor, sensor_maker, sensor_model, and sensor_serial_no.

argo_meta_param(): one row per file per parameter and columns file, n_param, parameter parameter_sensor, parameter_units, parameter_resolution, predeployment_calib_equation, predeployment_calib_coefficient, and predeployment_calib_comment.

Examples

with_argo_example_cache({
  argo_meta_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 7 × 5
#>   file                         n_config_param n_missions config_paramete… config_paramete…
#>   <chr>                                 <int>      <int>            <dbl> <chr>           
#> 1 csio/2900313/2900313_meta.nc              1          1             1500 CONFIG_ParkPres…
#> 2 csio/2900313/2900313_meta.nc              2          1             2000 CONFIG_ProfileP…
#> 3 csio/2900313/2900313_meta.nc              3          1               NA CONFIG_UpTime_h…
#> 4 csio/2900313/2900313_meta.nc              4          1               NA CONFIG_DownTime…
#> 5 csio/2900313/2900313_meta.nc              5          1              240 CONFIG_CycleTim…
#> 6 csio/2900313/2900313_meta.nc              6          1                1 CONFIG_Directio…
#> 7 csio/2900313/2900313_meta.nc              7          1                1 CONFIG_MeasureB…

with_argo_example_cache({
  argo_meta_missions("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 × 4
#>   file                         n_missions config_mission_number config_mission_…
#>   <chr>                             <int>                 <dbl> <chr>           
#> 1 csio/2900313/2900313_meta.nc          1                     1 ""              

with_argo_example_cache({
  argo_meta_trans_system("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 × 5
#>   file                         n_trans_system trans_system trans_system_id trans_frequency
#>   <chr>                                 <int> <chr>        <chr>           <chr>          
#> 1 csio/2900313/2900313_meta.nc              1 ARGOS        02528           1/45           

with_argo_example_cache({
  argo_meta_positioning_system("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 × 3
#>   file                         n_positioning_system positioning_system
#>   <chr>                                       <int> <chr>             
#> 1 csio/2900313/2900313_meta.nc                    1 ARGOS             

with_argo_example_cache({
  argo_meta_launch_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 7 × 4
#>   file                         n_launch_config… launch_config_p… launch_config_p…
#>   <chr>                                   <int> <chr>                       <dbl>
#> 1 csio/2900313/2900313_meta.nc                1 CONFIG_ParkPres…             1500
#> 2 csio/2900313/2900313_meta.nc                2 CONFIG_ProfileP…             2000
#> 3 csio/2900313/2900313_meta.nc                3 CONFIG_UpTime_h…               NA
#> 4 csio/2900313/2900313_meta.nc                4 CONFIG_DownTime…               NA
#> 5 csio/2900313/2900313_meta.nc                5 CONFIG_CycleTim…              240
#> 6 csio/2900313/2900313_meta.nc                6 CONFIG_Directio…                1
#> 7 csio/2900313/2900313_meta.nc                7 CONFIG_MeasureB…                1

with_argo_example_cache({
  argo_meta_sensor("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 × 6
#>   file                         n_sensor sensor sensor_maker sensor_model sensor_serial_no
#>   <chr>                           <int> <chr>  <chr>        <chr>        <chr>           
#> 1 csio/2900313/2900313_meta.nc        1 CTD_C… SBE          SBE41        0327            
#> 2 csio/2900313/2900313_meta.nc        2 CTD_T… SBE          SBE41        0327            
#> 3 csio/2900313/2900313_meta.nc        3 CTD_P… SBE          SBE41        0327            

with_argo_example_cache({
  argo_meta_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 × 10
#>   file       n_param parameter parameter_sensor parameter_units parameter_accur…
#>   <chr>        <int> <chr>     <chr>            <chr>           <chr>           
#> 1 csio/2900…       1 PSAL      CTD_CNDC         psu             0.005 psu       
#> 2 csio/2900…       2 TEMP      CTD_TEMP         degC (ITS-90)   0.002 degC      
#> 3 csio/2900…       3 PRES      CTD_PRES         dbar            1.5%            
#> # … with 4 more variables: parameter_resolution <chr>,
#> #   predeployment_calib_equation <chr>, predeployment_calib_coefficient <chr>,
#> #   predeployment_calib_comment <chr>