Load Argo float meta

Load Argo float meta

argo_meta_config_param(path, download = NULL, quiet = FALSE)

argo_meta_missions(path, download = NULL, quiet = FALSE)

argo_meta_trans_system(path, download = NULL, quiet = FALSE)

argo_meta_positioning_system(path, download = NULL, quiet = FALSE)

argo_meta_launch_config_param(path, download = NULL, quiet = FALSE)

argo_meta_sensor(path, download = NULL, quiet = FALSE)

argo_meta_param(path, download = NULL, quiet = FALSE)

Arguments

path
A path relative to the root directory of argo_mirror() or argo_cache_dir(). This value can also be a data.frame with a file column (e.g., a global index as returned by argo_global_meta() and others).

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

Value

A tibble::tibble().

Examples

with_argo_example_cache({
  argo_meta_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 7 x 4
#>   file             config_mission_num… config_parameter_name  config_parameter_…
#>   <chr>                          <int> <chr>                               <dbl>
#> 1 csio/2900313/29…                   1 CONFIG_ParkPressure_d…               1500
#> 2 csio/2900313/29…                   1 CONFIG_ProfilePressur…               2000
#> 3 csio/2900313/29…                   1 CONFIG_UpTime_hours                    NA
#> 4 csio/2900313/29…                   1 CONFIG_DownTime_hours                  NA
#> 5 csio/2900313/29…                   1 CONFIG_CycleTime_hours                240
#> 6 csio/2900313/29…                   1 CONFIG_Direction_NUMB…                  1
#> 7 csio/2900313/29…                   1 CONFIG_MeasureBattery…                  1

with_argo_example_cache({
  argo_meta_missions("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 x 3
#>   file                         config_mission_number config_mission_comment
#>   <chr>                                        <int> <chr>                 
#> 1 csio/2900313/2900313_meta.nc                     1 ""                    

with_argo_example_cache({
  argo_meta_trans_system("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 x 4
#>   file                         trans_system trans_system_id trans_frequency
#>   <chr>                        <chr>        <chr>           <chr>          
#> 1 csio/2900313/2900313_meta.nc ARGOS        02528           1/45           

with_argo_example_cache({
  argo_meta_positioning_system("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 1 x 2
#>   file                         positioning_system
#>   <chr>                        <chr>             
#> 1 csio/2900313/2900313_meta.nc ARGOS             

with_argo_example_cache({
  argo_meta_launch_config_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 7 x 3
#>   file                     launch_config_parameter_n… launch_config_parameter_v…
#>   <chr>                    <chr>                                           <dbl>
#> 1 csio/2900313/2900313_me… CONFIG_ParkPressure_dbar                         1500
#> 2 csio/2900313/2900313_me… CONFIG_ProfilePressure_db…                       2000
#> 3 csio/2900313/2900313_me… CONFIG_UpTime_hours                                NA
#> 4 csio/2900313/2900313_me… CONFIG_DownTime_hours                              NA
#> 5 csio/2900313/2900313_me… CONFIG_CycleTime_hours                            240
#> 6 csio/2900313/2900313_me… CONFIG_Direction_NUMBER                             1
#> 7 csio/2900313/2900313_me… CONFIG_MeasureBattery_LOG…                          1

with_argo_example_cache({
  argo_meta_sensor("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 x 5
#>   file                        sensor  sensor_maker sensor_model sensor_serial_no
#>   <chr>                       <chr>   <chr>        <chr>        <chr>           
#> 1 csio/2900313/2900313_meta.… CTD_CN… SBE          SBE41        0327            
#> 2 csio/2900313/2900313_meta.… CTD_TE… SBE          SBE41        0327            
#> 3 csio/2900313/2900313_meta.… CTD_PR… SBE          SBE41        0327            

with_argo_example_cache({
  argo_meta_param("dac/csio/2900313/2900313_meta.nc")
})

#> Extracting from 1 file

#> # A tibble: 3 x 9
#>   file  parameter parameter_sensor parameter_units parameter_accur…
#>   <chr> <chr>     <chr>            <chr>           <chr>           
#> 1 csio… PSAL      CTD_CNDC         psu             0.005 psu       
#> 2 csio… TEMP      CTD_TEMP         degC (ITS-90)   0.002 degC      
#> 3 csio… PRES      CTD_PRES         dbar            1.5%            
#> # … with 4 more variables: parameter_resolution <chr>,
#> #   predeployment_calib_equation <chr>, predeployment_calib_coefficient <chr>,
#> #   predeployment_calib_comment <chr>