Load Argo trajectories

Trajectory files contain (1) location measurements and (2) detailed information about each cycle. Use argo_traj_meas() output containing one row per location measurement; use argo_traj_cycle() for output containing one row per cycle. Finally, use argo_traj_history() to view quality control that has been applied.

argo_meta_config(path, download = NULL, quiet = FALSE)

argo_read_meta_config(file, vars = NULL)

Arguments

path
A path relative to the root directory of argo_mirror().

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

file
A previously downloaded Argo NetCDF file.

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

Value

A tibble::tibble().

Examples

with_argo_example_cache({
  argo_meta_config("dac/csio/2900313/2900313_meta.nc")
})

#> # A tibble: 7 x 4
#>   float      config_mission_numb… config_parameter_name     config_parameter_va…
#>   <chr>                     <int> <chr>                                    <dbl>
#> 1 csio/2900…                    1 CONFIG_ParkPressure_dbar                  1500
#> 2 csio/2900…                    1 CONFIG_ProfilePressure_d…                 2000
#> 3 csio/2900…                    1 CONFIG_UpTime_hours                         NA
#> 4 csio/2900…                    1 CONFIG_DownTime_hours                       NA
#> 5 csio/2900…                    1 CONFIG_CycleTime_hours                     240
#> 6 csio/2900…                    1 CONFIG_Direction_NUMBER                      1
#> 7 csio/2900…                    1 CONFIG_MeasureBattery_LO…                    1

meta_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_meta.nc",
  package = "argodata"
)

argo_read_meta_config(meta_file)

#> # A tibble: 7 x 4
#>   float   CONFIG_MISSION_NU… CONFIG_PARAMETER_NAME             CONFIG_PARAMETER…
#>   <chr>                <int> <chr>                                         <dbl>
#> 1 csio/2…                  1 "CONFIG_ParkPressure_dbar       …              1500
#> 2 csio/2…                  1 "CONFIG_ProfilePressure_dbar    …              2000
#> 3 csio/2…                  1 "CONFIG_UpTime_hours            …                NA
#> 4 csio/2…                  1 "CONFIG_DownTime_hours          …                NA
#> 5 csio/2…                  1 "CONFIG_CycleTime_hours         …               240
#> 6 csio/2…                  1 "CONFIG_Direction_NUMBER        …                 1
#> 7 csio/2…                  1 "CONFIG_MeasureBattery_LOGICAL  …                 1