Load Argo profiles

Load Argo profiles

argo_prof_levels(path, vars = NULL, download = NULL, quiet = FALSE)

argo_prof_prof(path, vars = NULL, download = NULL, quiet = FALSE)

argo_prof_history(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_prof_levels(file, vars = NULL, meta = NULL)

argo_read_prof_prof(file, vars = NULL)

argo_read_prof_history(file, vars = NULL, meta = NULL)

Arguments

path
A path relative to the root directory of argo_mirror() or argo_cache_dir(). This value can also be a data.frame with a file column (e.g., a global index as returned by argo_global_meta() and others).

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

file
A previously downloaded Argo NetCDF file.

meta
A character vector of meta variable names for each profile. These can be found using argo_vars().

Value

A tibble::tibble().

Level Variables

Some of these variables may be missing depending on the sensors that are available for a given float.

pres (decibar, float): Sea water pressure, equals 0 at sea-level

pres_qc (char): quality flag

pres_adjusted (decibar, float): Sea water pressure, equals 0 at sea-level

pres_adjusted_qc (char): quality flag

pres_adjusted_error (decibar, float): Contains the error on the adjusted values as determined by the delayed mode QC process

temp (degree_Celsius, float): Sea temperature in-situ ITS-90 scale

temp_qc (char): quality flag

temp_adjusted (degree_Celsius, float): Sea temperature in-situ ITS-90 scale

temp_adjusted_qc (char): quality flag

temp_adjusted_error (degree_Celsius, float): Contains the error on the adjusted values as determined by the delayed mode QC process

psal (psu, float): Practical salinity

psal_qc (char): quality flag

psal_adjusted (psu, float): Practical salinity

psal_adjusted_qc (char): quality flag

psal_adjusted_error (psu, float): Contains the error on the adjusted values as determined by the delayed mode QC process

Profile Variables

cycle_number (int): Float cycle number

direction (char): Direction of the station profiles

data_mode (char): Delayed mode or real time data

date (days since 1950-01-01 00:00:00 UTC, double): Julian day (UTC) of the station relative to REFERENCE_DATE_TIME

date_qc (char): Quality on date and time

date_location (days since 1950-01-01 00:00:00 UTC, double): Julian day (UTC) of the location relative to REFERENCE_DATE_TIME

latitude (degree_north, double): Latitude of the station, best estimate

longitude (degree_east, double): Longitude of the station, best estimate

position_qc (char): Quality on position (latitude and longitude)

profile_pres_qc (char): Global quality flag of PRES profile

profile_temp_qc (char): Global quality flag of TEMP profile

profile_psal_qc (char): Global quality flag of PSAL profile

config_mission_number (int): Unique number denoting the missions performed by the float

History variables

history_start_pres (decibar, float): Start pressure action applied on

history_stop_pres (decibar, float): Stop pressure action applied on

history_previous_value (float): Parameter/Flag previous value before action

history_institution (char): Institution which performed action

history_step (char): Step in data processing

history_software (char): Name of software which performed action

history_software_release (char): Version/release of software which performed action

history_reference (char): Reference of database

history_date (char): Date the history record was created

history_action (char): Action performed on data

history_parameter (char): Station parameter action is performed on

history_qctest (char): Documentation of tests performed, tests failed (in hex form)

Examples

with_argo_example_cache({
  argo_prof_levels("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> # A tibble: 70 x 18
#>    file  cycle_number date                  pres pres_qc pres_adjusted
#>    <chr>        <int> <dttm>               <dbl> <chr>           <dbl>
#>  1 csio…            0 2003-08-13 08:47:42   9.80 1                9.80
#>  2 csio…            0 2003-08-13 08:47:42  20.1  1               20.1 
#>  3 csio…            0 2003-08-13 08:47:42  29.9  1               29.9 
#>  4 csio…            0 2003-08-13 08:47:42  39.7  1               39.7 
#>  5 csio…            0 2003-08-13 08:47:42  49.9  1               49.9 
#>  6 csio…            0 2003-08-13 08:47:42  60.3  1               60.3 
#>  7 csio…            0 2003-08-13 08:47:42  69.7  1               69.7 
#>  8 csio…            0 2003-08-13 08:47:42  80.3  1               80.3 
#>  9 csio…            0 2003-08-13 08:47:42  90.2  1               90.2 
#> 10 csio…            0 2003-08-13 08:47:42 100.   1              100.  
#> # … with 60 more rows, and 12 more variables: pres_adjusted_qc <chr>,
#> #   pres_adjusted_error <dbl>, temp <dbl>, temp_qc <chr>, temp_adjusted <dbl>,
#> #   temp_adjusted_qc <chr>, temp_adjusted_error <dbl>, psal <dbl>,
#> #   psal_qc <chr>, psal_adjusted <dbl>, psal_adjusted_qc <chr>,
#> #   psal_adjusted_error <dbl>

with_argo_example_cache({
  argo_prof_prof("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> # A tibble: 1 x 14
#>   file  cycle_number direction data_mode date                date_qc            
#>   <chr>        <int> <chr>     <chr>     <dttm>              <dttm>             
#> 1 csio…            0 A         D         2003-08-13 08:47:42 NA                 
#> # … with 8 more variables: date_location <dttm>, latitude <dbl>,
#> #   longitude <dbl>, position_qc <chr>, profile_pres_qc <chr>,
#> #   profile_temp_qc <chr>, profile_psal_qc <chr>, config_mission_number <int>

with_argo_example_cache({
  argo_prof_history("dac/csio/2900313/profiles/D2900313_000.nc")
})

#> # A tibble: 3 x 15
#>   file  cycle_number date                history_start_p… history_stop_pr…
#>   <chr>        <int> <dttm>                         <dbl>            <dbl>
#> 1 csio…            0 2003-08-13 08:47:42               NA               NA
#> 2 csio…            0 2003-08-13 08:47:42               NA               NA
#> 3 csio…            0 2003-08-13 08:47:42               NA               NA
#> # … with 10 more variables: history_previous_value <dbl>,
#> #   history_institution <chr>, history_step <chr>, history_software <chr>,
#> #   history_software_release <chr>, history_reference <chr>,
#> #   history_date <chr>, history_action <chr>, history_parameter <chr>,
#> #   history_qctest <chr>

prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_prof_levels(prof_file)

#> # A tibble: 70 x 28
#>    CYCLE_NUMBER DIRECTION DATA_MODE   JULD JULD_QC JULD_LOCATION LATITUDE
#>           <int> <chr>     <chr>      <dbl> <chr>           <dbl>    <dbl>
#>  1            0 A         D         19582. 1              19582.     23.2
#>  2            0 A         D         19582. 1              19582.     23.2
#>  3            0 A         D         19582. 1              19582.     23.2
#>  4            0 A         D         19582. 1              19582.     23.2
#>  5            0 A         D         19582. 1              19582.     23.2
#>  6            0 A         D         19582. 1              19582.     23.2
#>  7            0 A         D         19582. 1              19582.     23.2
#>  8            0 A         D         19582. 1              19582.     23.2
#>  9            0 A         D         19582. 1              19582.     23.2
#> 10            0 A         D         19582. 1              19582.     23.2
#> # … with 60 more rows, and 21 more variables: LONGITUDE <dbl>,
#> #   POSITION_QC <chr>, PROFILE_PRES_QC <chr>, PROFILE_TEMP_QC <chr>,
#> #   PROFILE_PSAL_QC <chr>, CONFIG_MISSION_NUMBER <int>, PRES <dbl>,
#> #   PRES_QC <chr>, PRES_ADJUSTED <dbl>, PRES_ADJUSTED_QC <chr>,
#> #   PRES_ADJUSTED_ERROR <dbl>, TEMP <dbl>, TEMP_QC <chr>, TEMP_ADJUSTED <dbl>,
#> #   TEMP_ADJUSTED_QC <chr>, TEMP_ADJUSTED_ERROR <dbl>, PSAL <dbl>,
#> #   PSAL_QC <chr>, PSAL_ADJUSTED <dbl>, PSAL_ADJUSTED_QC <chr>,
#> #   PSAL_ADJUSTED_ERROR <dbl>

argo_read_prof_prof(prof_file)

#> # A tibble: 1 x 13
#>   CYCLE_NUMBER DIRECTION DATA_MODE   JULD JULD_QC JULD_LOCATION LATITUDE
#>          <int> <chr>     <chr>      <dbl> <chr>           <dbl>    <dbl>
#> 1            0 A         D         19582. 1              19582.     23.2
#> # … with 6 more variables: LONGITUDE <dbl>, POSITION_QC <chr>,
#> #   PROFILE_PRES_QC <chr>, PROFILE_TEMP_QC <chr>, PROFILE_PSAL_QC <chr>,
#> #   CONFIG_MISSION_NUMBER <int>

argo_read_prof_history(prof_file)

#> # A tibble: 3 x 25
#>   CYCLE_NUMBER DIRECTION DATA_MODE   JULD JULD_QC JULD_LOCATION LATITUDE
#>          <int> <chr>     <chr>      <dbl> <chr>           <dbl>    <dbl>
#> 1            0 A         D         19582. 1              19582.     23.2
#> 2            0 A         D         19582. 1              19582.     23.2
#> 3            0 A         D         19582. 1              19582.     23.2
#> # … with 18 more variables: LONGITUDE <dbl>, POSITION_QC <chr>,
#> #   PROFILE_PRES_QC <chr>, PROFILE_TEMP_QC <chr>, PROFILE_PSAL_QC <chr>,
#> #   CONFIG_MISSION_NUMBER <int>, HISTORY_START_PRES <dbl>,
#> #   HISTORY_STOP_PRES <dbl>, HISTORY_PREVIOUS_VALUE <dbl>,
#> #   HISTORY_INSTITUTION <chr>, HISTORY_STEP <chr>, HISTORY_SOFTWARE <chr>,
#> #   HISTORY_SOFTWARE_RELEASE <chr>, HISTORY_REFERENCE <chr>,
#> #   HISTORY_DATE <chr>, HISTORY_ACTION <chr>, HISTORY_PARAMETER <chr>,
#> #   HISTORY_QCTEST <chr>