Read Argo float meta

Use argo_read_meta_*() functions to extract meta information from a previously-downloaded Argo NetCDF file.

Use argo_read_tech_tech_param() to extract technical specifications from a previously-downloaded Argo NetCDF file.

argo_read_meta_config_param(file, vars = NULL, quiet = FALSE)

argo_read_meta_missions(file, vars = NULL, quiet = FALSE)

argo_read_meta_trans_system(file, vars = NULL, quiet = FALSE)

argo_read_meta_positioning_system(file, vars = NULL, quiet = FALSE)

argo_read_meta_launch_config_param(file, vars = NULL, quiet = TRUE)

argo_read_meta_sensor(file, vars = NULL, quiet = quiet)

argo_read_meta_param(file, vars = NULL, quiet = quiet)

argo_read_tech_tech_param(file, vars = NULL, quiet = FALSE)

Arguments

file

A previously downloaded Argo NetCDF file (e.g., using argo_download()).

vars

A vector of variable names to include. Explicitly specifying vars can lead to much faster read times when reading many files.

quiet

Use FALSE to stop for malformed files, NA to silently warn for malformed files, or TRUE to silently ignore read errors when possible.

Value

A tibble::tibble() with

argo_meta_config_param(): one row per configuration parameter and columns n_config_param, N_MISSIONS, CONFIG_PARAMETER_VALUE, and CONFIG_PARAMETER_NAME.

argo_meta_missions(): one row per mission and columns N_MISSIONS, CONFIG_MISSION_NUMBER, and CONFIG_MISSION_COMMENT.

argo_meta_trans_system(): one row per transmission system and columns N_TRANS_SYSTEM, TRANS_SYSTEM, TRANS_SYSTEM_ID, and TRANS_FREQUENCY.

argo_meta_positioning_system(): one row per positioning system and columns N_POSITIONING_SYSTEM, and POSITIONING_SYSTEM.

argo_meta_launch_config_param(): one row per launch configuration parameter and columns N_LAUNCH_CONFIG_PARAM, LAUNCH_CONFIG_PARAMETER_NAME, and LAUNCH_CONFIG_PARAMETER_VALUE.

argo_meta_sensor(): one row per sensor and columns N_SENSOR, SENSOR, SENSOR_MAKER, SENSOR_MODEL, and sensor_serial_no.

argo_meta_param(): one row per parameter and columns N_PARAM, PARAMETER PARAMETER_SENSOR, PARAMETER_UNITS, PARAMETER_RESOLUTION, PREDEPLOYMENT_CALIB_EQUATION, PREDEPLOYMENT_CALIB_COEFFICIENT, and PREDEPLOYMENT_CALIB_COMMENT.

A tibble::tibble() with one row per technical parameter and columns N_TECH_PARAM, TECHNICAL_PARAMETER_NAME, TECHNICAL_PARAMETER_VALUE, and CYCLE_NUMBER.

Examples

meta_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_meta.nc",
  package = "argodata"
)

argo_read_meta_config_param(meta_file)
#> # A tibble: 7 × 4
#>   N_CONFIG_PARAM N_MISSIONS CONFIG_PARAMETER_VALUE CONFIG_PARAMETER_NAME        
#>            <int>      <int>                  <dbl> <chr>                        
#> 1              1          1                   1500 "CONFIG_ParkPressure_dbar   …
#> 2              2          1                   2000 "CONFIG_ProfilePressure_dbar…
#> 3              3          1                     NA "CONFIG_UpTime_hours        …
#> 4              4          1                     NA "CONFIG_DownTime_hours      …
#> 5              5          1                    240 "CONFIG_CycleTime_hours     …
#> 6              6          1                      1 "CONFIG_Direction_NUMBER    …
#> 7              7          1                      1 "CONFIG_MeasureBattery_LOGIC…
argo_read_meta_missions(meta_file)
#> # A tibble: 1 × 3
#>   N_MISSIONS CONFIG_MISSION_NUMBER CONFIG_MISSION_COMMENT                       
#> *      <int>                 <dbl> <chr>                                        
#> 1          1                     1 "                                           …
argo_read_meta_trans_system(meta_file)
#> # A tibble: 1 × 4
#>   N_TRANS_SYSTEM TRANS_SYSTEM       TRANS_SYSTEM_ID             TRANS_FREQUENCY 
#> *          <int> <chr>              <chr>                       <chr>           
#> 1              1 "ARGOS           " "02528                    … "1/45          …
argo_read_meta_positioning_system(meta_file)
#> # A tibble: 1 × 2
#>   N_POSITIONING_SYSTEM POSITIONING_SYSTEM
#> *                <int> <chr>             
#> 1                    1 "ARGOS   "        
argo_read_meta_launch_config_param(meta_file)
#> # A tibble: 7 × 3
#>   N_LAUNCH_CONFIG_PARAM LAUNCH_CONFIG_PARAMETER_NAME         LAUNCH_CONFIG_PARA…
#> *                 <int> <chr>                                              <dbl>
#> 1                     1 "CONFIG_ParkPressure_dbar          …                1500
#> 2                     2 "CONFIG_ProfilePressure_dbar       …                2000
#> 3                     3 "CONFIG_UpTime_hours               …                  NA
#> 4                     4 "CONFIG_DownTime_hours             …                  NA
#> 5                     5 "CONFIG_CycleTime_hours            …                 240
#> 6                     6 "CONFIG_Direction_NUMBER           …                   1
#> 7                     7 "CONFIG_MeasureBattery_LOGICAL     …                   1
argo_read_meta_sensor(meta_file)
#> # A tibble: 3 × 5
#>   N_SENSOR SENSOR     SENSOR_MAKER         SENSOR_MODEL         SENSOR_SERIAL_NO
#> *    <int> <chr>      <chr>                <chr>                <chr>           
#> 1        1 "CTD_CNDC… "SBE               … "SBE41             … "0327          …
#> 2        2 "CTD_TEMP… "SBE               … "SBE41             … "0327          …
#> 3        3 "CTD_PRES… "SBE               … "SBE41             … "0327          …
argo_read_meta_param(meta_file)
#> # A tibble: 3 × 9
#>   N_PARAM PARAMETER      PARAMETER_SENSOR     PARAMETER_UNITS  PARAMETER_ACCURA…
#> *   <int> <chr>          <chr>                <chr>            <chr>            
#> 1       1 "PSAL        … "CTD_CNDC          … "psu           … "0.005 psu      …
#> 2       2 "TEMP        … "CTD_TEMP          … "degC (ITS-90) … "0.002 degC     …
#> 3       3 "PRES        … "CTD_PRES          … "dbar          … "1.5%           …
#> # … with 4 more variables: PARAMETER_RESOLUTION <chr>,
#> #   PREDEPLOYMENT_CALIB_EQUATION <chr>, PREDEPLOYMENT_CALIB_COEFFICIENT <chr>,
#> #   PREDEPLOYMENT_CALIB_COMMENT <chr>

tech_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_tech.nc",
  package = "argodata"
)

argo_read_tech_tech_param(tech_file)
#> # A tibble: 96 × 4
#>    N_TECH_PARAM TECHNICAL_PARAMETER_NAME  TECHNICAL_PARAMETER_VALUE CYCLE_NUMBER
#>  *        <int> <chr>                     <chr>                            <dbl>
#>  1            1 "NUMBER_ValveActionsAtSu… "150                    …            0
#>  2            2 "CLOCK_InitialStabilizat… " 0.2                   …            0
#>  3            3 "NUMBER_PumpActionsDurin… "0                      …            0
#>  4            4 "NUMBER_RepositionsDurin… "0                      …            0
#>  5            5 "NUMBER_PumpActionsDurin… "14                     …            0
#>  6            6 "NUMBER_PumpActionsAtSur… "2                      …            0
#>  7            7 "PRES_SurfaceOffsetCorre… "1.025                  …            0
#>  8            8 "PRESSURE_InternalVacuum… "740.0                  …            0
#>  9            9 "NUMBER_AscentArgosMessa… "14                     …            0
#> 10           10 "NUMBER_AscentSamples_CO… "71                     …            0
#> # … with 86 more rows