Read NetCDF variable metadata

Read NetCDF variable metadata

argo_read_vars(file, vars = NULL)

Arguments

file
A previously downloaded Argo NetCDF file.

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

Value

A tibble::tibble() with one row per variable and columns name, longname, units, prec, ndims, size, and dim. Columns size and dim can be unnested (e.g., using tidyr::unnest()) to produce a tibble with one row per variable per dimension.

Examples

prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_vars(prof_file)

#> # A tibble: 64 x 18
#>    name  longname units prec  ndims size  dim   att_long_name att_conventions
#>    <chr> <chr>    <chr> <chr> <int> <lis> <lis> <chr>         <chr>          
#>  1 DATA… Data ty… ""    char      1 <int… <chr… Data type     Argo reference…
#>  2 FORM… File fo… ""    char      1 <int… <chr… File format … NA             
#>  3 HAND… Data ha… ""    char      1 <int… <chr… Data handboo… NA             
#>  4 REFE… Date of… ""    char      1 <int… <chr… Date of refe… YYYYMMDDHHMISS 
#>  5 DATE… Date of… ""    char      1 <int… <chr… Date of file… YYYYMMDDHHMISS 
#>  6 DATE… Date of… ""    char      1 <int… <chr… Date of upda… YYYYMMDDHHMISS 
#>  7 PLAT… Float u… ""    char      2 <int… <chr… Float unique… WMO float iden…
#>  8 PLAT… Type of… ""    char      2 <int… <chr… Type of float Argo reference…
#>  9 PROJ… Name of… ""    char      2 <int… <chr… Name of the … NA             
#> 10 PI_N… Name of… ""    char      2 <int… <chr… Name of the … NA             
#> # … with 54 more rows, and 9 more variables: att_standard_name <chr>,
#> #   att_units <chr>, att_resolution <dbl>, att_axis <chr>, att_valid_min <dbl>,
#> #   att_valid_max <dbl>, att_C_format <chr>, att_FORTRAN_format <chr>,
#> #   att_comment <chr>