Load Argo trajectories

Trajectory files contain (1) location measurements and (2) detailed information about each cycle. Use argo_traj_meas() output containing one row per location measurement; use argo_traj_cycle() for output containing one row per cycle. Finally, use argo_traj_history() to view quality control that has been applied.

argo_traj_meas(path, vars = NULL, download = NULL, quiet = FALSE)

argo_traj_cycle(path, vars = NULL, download = NULL, quiet = FALSE)

argo_traj_history(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_traj_meas(file, vars = NULL)

argo_read_traj_cycle(file, vars = NULL)

argo_read_traj_history(file, vars = NULL)

Arguments

path
A path relative to the root directory of argo_mirror().

vars
A vector of variable names to include. The ordering of the variables is not guaranteed, and variables that do not exist are ignored. For nc_read_*() and read_*() variants, these are the raw variable names (all caps, without "date" substituted for "juld").

download
A logical vector indicating whether or not a file should be downloaded. Defaults to the value of argo_should_download(), which is TRUE for files that do not exist in the cache or global files that were cached more than 24 hours ago.

quiet
Use FALSE to show which files are downloaded and for more verbose error messages.

file
A previously downloaded Argo NetCDF file.

Value

A tibble::tibble().

Measurement Variables

date (days since 1950-01-01 00:00:00 UTC, double): Julian day (UTC) of each measurement relative to REFERENCE_DATE_TIME

date_status (char): Status of the date and time

date_qc (char): Quality on date and time

date_adjusted (days since 1950-01-01 00:00:00 UTC, double): Adjusted julian day (UTC) of each measurement relative to REFERENCE_DATE_TIME

date_adjusted_status (char): Status of the JULD_ADJUSTED date

date_adjusted_qc (char): Quality on adjusted date and time

latitude (degree_north, double): Latitude of each location

longitude (degree_east, double): Longitude of each location

position_accuracy (char): Estimated accuracy in latitude and longitude

position_qc (char): Quality on position

cycle_number (int): Float cycle number of the measurement

cycle_number_adjusted (int): Adjusted float cycle number of the measurement

measurement_code (int): Flag referring to a measurement event in the cycle

pres (decibar, float): Sea water pressure, equals 0 at sea-level

pres_qc (char): quality flag

pres_adjusted (decibar, float): Sea water pressure, equals 0 at sea-level

pres_adjusted_qc (char): quality flag

pres_adjusted_error (decibar, float): Contains the error on the adjusted values as determined by the delayed mode QC process

temp (degree_Celsius, float): Sea temperature in-situ ITS-90 scale

temp_qc (char): quality flag

temp_adjusted (degree_Celsius, float): Sea temperature in-situ ITS-90 scale

temp_adjusted_qc (char): quality flag

temp_adjusted_error (degree_Celsius, float): Contains the error on the adjusted values as determined by the delayed mode QC process

psal (psu, float): Practical salinity

psal_qc (char): quality flag

psal_adjusted (psu, float): Practical salinity

psal_adjusted_qc (char): quality flag

psal_adjusted_error (psu, float): Contains the error on the adjusted values as determined by the delayed mode QC process

axes_error_ellipse_major (meters, float): Major axis of error ellipse from positioning system

axes_error_ellipse_minor (meters, float): Minor axis of error ellipse from positioning system

axes_error_ellipse_angle (Degrees (from North when heading East), float): Angle of error ellipse from positioning system

satellite_name (char): Satellite name from positioning system

Cycle Variables

date_descent_start (days since 1950-01-01 00:00:00 UTC, double): Descent start date of the cycle

date_descent_start_status (char): Status of descent start date of the cycle

date_first_stabilization (days since 1950-01-01 00:00:00 UTC, double): Time when a float first becomes water-neutral

date_first_stabilization_status (char): Status of time when a float first becomes water-neutral

date_descent_end (days since 1950-01-01 00:00:00 UTC, double): Descent end date of the cycle

date_descent_end_status (char): Status of descent end date of the cycle

date_park_start (days since 1950-01-01 00:00:00 UTC, double): Drift start date of the cycle

date_park_start_status (char): Status of drift start date of the cycle

date_park_end (days since 1950-01-01 00:00:00 UTC, double): Drift end date of the cycle

date_park_end_status (char): Status of drift end date of the cycle

date_deep_descent_end (days since 1950-01-01 00:00:00 UTC, double): Deep descent end date of the cycle

date_deep_descent_end_status (char): Status of deep descent end date of the cycle

date_deep_park_start (days since 1950-01-01 00:00:00 UTC, double): Deep park start date of the cycle

date_deep_park_start_status (char): Status of deep park start date of the cycle

date_ascent_start (days since 1950-01-01 00:00:00 UTC, double): Start date of the ascent to the surface

date_ascent_start_status (char): Status of start date of the ascent to the surface

date_deep_ascent_start (days since 1950-01-01 00:00:00 UTC, double): Deep ascent start date of the cycle

date_deep_ascent_start_status (char): Status of deep ascent start date of the cycle

date_ascent_end (days since 1950-01-01 00:00:00 UTC, double): End date of ascent to the surface

date_ascent_end_status (char): Status of end date of ascent to the surface

date_transmission_start (days since 1950-01-01 00:00:00 UTC, double): Start date of transmission

date_transmission_start_status (char): Status of start date of transmission

date_first_message (days since 1950-01-01 00:00:00 UTC, double): Date of earliest float message received

date_first_message_status (char): Status of date of earliest float message received

date_first_location (days since 1950-01-01 00:00:00 UTC, double): Date of earliest location

date_first_location_status (char): Status of date of earliest location

date_last_location (days since 1950-01-01 00:00:00 UTC, double): Date of latest location

date_last_location_status (char): Status of date of latest location

date_last_message (days since 1950-01-01 00:00:00 UTC, double): Date of latest float message received

date_last_message_status (char): Status of date of latest float message received

date_transmission_end (days since 1950-01-01 00:00:00 UTC, double): Transmission end date

date_transmission_end_status (char): Status of transmission end date

clock_offset (days, double): Time of float clock drift

grounded (days, char): Did the profiler touch the ground for that cycle?

representative_park_pressure (decibar, float): Best pressure value during park phase

representative_park_pressure_status (char): Status of best pressure value during park phase

config_mission_number (int): Unique number denoting the missions performed by the float

cycle_number_index (int): Cycle number that corresponds to the current index

cycle_number_index_adjusted (int): Adjusted cycle number that corresponds to the current index

data_mode (char): Delayed mode or real time data

History Variables

history_previous_value (float): Parameter/Flag previous value before action

history_index_dimension (char): Name of dimension to which HISTORY_START_INDEX and HISTORY_STOP_INDEX correspond

history_start_index (int): Start index action applied on

history_stop_index (int): Stop index action applied on

history_institution (char): Institution which performed action

history_step (char): Step in data processing

history_software (char): Name of software which performed action

history_software_release (char): Version/release of software which performed action

history_reference (char): Reference of database

history_date (char): Date the history record was created

history_action (char): Action performed on data

history_parameter (char): Station parameter action is performed on

history_qctest (char): Documentation of tests performed, tests failed (in hex form)

Examples

with_argo_example_cache({
  argo_traj_meas("dac/csio/2900313/2900313_Rtraj.nc")
})

#> # A tibble: 61 x 33
#>    float date                date_status         date_qc            
#>    <chr> <dttm>              <dttm>              <dttm>             
#>  1 csio… 2003-08-04 01:40:00 NA                  NA                 
#>  2 csio… 2003-08-04 01:40:00 NA                  NA                 
#>  3 csio… 2003-08-13 08:52:56 NA                  NA                 
#>  4 csio… 2003-08-13 12:09:46 NA                  NA                 
#>  5 csio… 2003-08-13 13:47:15 NA                  NA                 
#>  6 csio… 2003-08-13 16:59:59 NA                  NA                 
#>  7 csio… 2003-08-13 18:41:17 NA                  NA                 
#>  8 csio… 2003-08-13 18:46:50 NA                  NA                 
#>  9 csio… 2003-08-23 08:15:49 NA                  NA                 
#> 10 csio… 2003-08-23 08:19:09 NA                  NA                 
#> # … with 51 more rows, and 29 more variables: date_adjusted <dttm>,
#> #   date_adjusted_status <dttm>, date_adjusted_qc <dttm>, latitude <dbl>,
#> #   longitude <dbl>, position_accuracy <chr>, position_qc <chr>,
#> #   cycle_number <int>, cycle_number_adjusted <int>, measurement_code <int>,
#> #   pres <dbl>, pres_qc <chr>, pres_adjusted <dbl>, pres_adjusted_qc <chr>,
#> #   pres_adjusted_error <dbl>, temp <dbl>, temp_qc <chr>, temp_adjusted <dbl>,
#> #   temp_adjusted_qc <chr>, temp_adjusted_error <dbl>, psal <dbl>,
#> #   psal_qc <chr>, psal_adjusted <dbl>, psal_adjusted_qc <chr>,
#> #   psal_adjusted_error <dbl>, axes_error_ellipse_major <dbl>,
#> #   axes_error_ellipse_minor <dbl>, axes_error_ellipse_angle <dbl>,
#> #   satellite_name <chr>

with_argo_example_cache({
  argo_traj_cycle("dac/csio/2900313/2900313_Rtraj.nc")
})

#> # A tibble: 8 x 41
#>   float date_descent_start  date_descent_start… date_first_stabili…
#>   <chr> <dttm>              <dttm>              <dttm>             
#> 1 csio… 2003-08-04 04:34:33 NA                  NA                 
#> 2 csio… 2003-08-22 04:49:08 NA                  NA                 
#> 3 csio… 2003-09-01 04:52:03 NA                  NA                 
#> 4 csio… 2003-09-11 05:09:28 NA                  NA                 
#> 5 csio… 2003-09-21 09:53:54 NA                  NA                 
#> 6 csio… 2003-10-01 05:15:36 NA                  NA                 
#> 7 csio… 2003-10-11 04:04:28 NA                  NA                 
#> 8 csio… 2003-10-21 23:39:28 NA                  NA                 
#> # … with 37 more variables: date_first_stabilization_status <dttm>,
#> #   date_descent_end <dttm>, date_descent_end_status <dttm>,
#> #   date_park_start <dttm>, date_park_start_status <dttm>,
#> #   date_park_end <dttm>, date_park_end_status <dttm>,
#> #   date_deep_descent_end <dttm>, date_deep_descent_end_status <dttm>,
#> #   date_deep_park_start <dttm>, date_deep_park_start_status <dttm>,
#> #   date_ascent_start <dttm>, date_ascent_start_status <dttm>,
#> #   date_deep_ascent_start <dttm>, date_deep_ascent_start_status <dttm>,
#> #   date_ascent_end <dttm>, date_ascent_end_status <dttm>,
#> #   date_transmission_start <dttm>, date_transmission_start_status <dttm>,
#> #   date_first_message <dttm>, date_first_message_status <dttm>,
#> #   date_first_location <dttm>, date_first_location_status <dttm>,
#> #   date_last_location <dttm>, date_last_location_status <dttm>,
#> #   date_last_message <dttm>, date_last_message_status <dttm>,
#> #   date_transmission_end <dttm>, date_transmission_end_status <dttm>,
#> #   clock_offset <dbl>, grounded <chr>, representative_park_pressure <dbl>,
#> #   representative_park_pressure_status <chr>, config_mission_number <int>,
#> #   cycle_number_index <int>, cycle_number_index_adjusted <int>,
#> #   data_mode <chr>

with_argo_example_cache({
  argo_traj_history("dac/csio/2900313/2900313_Rtraj.nc")
})

#> # A tibble: 1 x 14
#>   float history_previou… history_index_d… history_start_i… history_stop_in…
#>   <chr>            <dbl> <chr>                       <int>            <int>
#> 1 csio…               NA " "                            NA               NA
#> # … with 9 more variables: history_institution <chr>, history_step <chr>,
#> #   history_software <chr>, history_software_release <chr>,
#> #   history_reference <chr>, history_date <chr>, history_action <chr>,
#> #   history_parameter <chr>, history_qctest <chr>

traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)

argo_read_traj_meas(traj_file)

#> # A tibble: 61 x 33
#>    float   JULD JULD_STATUS JULD_QC JULD_ADJUSTED JULD_ADJUSTED_S…
#>    <chr>  <dbl> <chr>       <chr>           <dbl> <chr>           
#>  1 csio… 19573. 0           0                  NA " "             
#>  2 csio… 19573. 4           1                  NA " "             
#>  3 csio… 19582. 4           1                  NA " "             
#>  4 csio… 19583. 4           1                  NA " "             
#>  5 csio… 19583. 4           1                  NA " "             
#>  6 csio… 19583. 4           1                  NA " "             
#>  7 csio… 19583. 4           1                  NA " "             
#>  8 csio… 19583. 4           1                  NA " "             
#>  9 csio… 19592. 4           1                  NA " "             
#> 10 csio… 19592. 4           1                  NA " "             
#> # … with 51 more rows, and 27 more variables: JULD_ADJUSTED_QC <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, POSITION_ACCURACY <chr>,
#> #   POSITION_QC <chr>, CYCLE_NUMBER <int>, CYCLE_NUMBER_ADJUSTED <int>,
#> #   MEASUREMENT_CODE <int>, PRES <dbl>, PRES_QC <chr>, PRES_ADJUSTED <dbl>,
#> #   PRES_ADJUSTED_QC <chr>, PRES_ADJUSTED_ERROR <dbl>, TEMP <dbl>,
#> #   TEMP_QC <chr>, TEMP_ADJUSTED <dbl>, TEMP_ADJUSTED_QC <chr>,
#> #   TEMP_ADJUSTED_ERROR <dbl>, PSAL <dbl>, PSAL_QC <chr>, PSAL_ADJUSTED <dbl>,
#> #   PSAL_ADJUSTED_QC <chr>, PSAL_ADJUSTED_ERROR <dbl>,
#> #   AXES_ERROR_ELLIPSE_MAJOR <dbl>, AXES_ERROR_ELLIPSE_MINOR <dbl>,
#> #   AXES_ERROR_ELLIPSE_ANGLE <dbl>, SATELLITE_NAME <chr>

argo_read_traj_cycle(traj_file)

#> # A tibble: 8 x 41
#>   float JULD_DESCENT_ST… JULD_DESCENT_ST… JULD_FIRST_STAB… JULD_FIRST_STAB…
#>   <chr>            <dbl> <chr>                       <dbl> <chr>           
#> 1 csio…           19573. 2                              NA 9               
#> 2 csio…           19591. 2                              NA 9               
#> 3 csio…           19601. 2                              NA 9               
#> 4 csio…           19611. 2                              NA 9               
#> 5 csio…           19621. 2                              NA 9               
#> 6 csio…           19631. 2                              NA 9               
#> 7 csio…           19641. 2                              NA 9               
#> 8 csio…           19652. 0                              NA 9               
#> # … with 36 more variables: JULD_DESCENT_END <dbl>,
#> #   JULD_DESCENT_END_STATUS <chr>, JULD_PARK_START <dbl>,
#> #   JULD_PARK_START_STATUS <chr>, JULD_PARK_END <dbl>,
#> #   JULD_PARK_END_STATUS <chr>, JULD_DEEP_DESCENT_END <dbl>,
#> #   JULD_DEEP_DESCENT_END_STATUS <chr>, JULD_DEEP_PARK_START <dbl>,
#> #   JULD_DEEP_PARK_START_STATUS <chr>, JULD_ASCENT_START <dbl>,
#> #   JULD_ASCENT_START_STATUS <chr>, JULD_DEEP_ASCENT_START <dbl>,
#> #   JULD_DEEP_ASCENT_START_STATUS <chr>, JULD_ASCENT_END <dbl>,
#> #   JULD_ASCENT_END_STATUS <chr>, JULD_TRANSMISSION_START <dbl>,
#> #   JULD_TRANSMISSION_START_STATUS <chr>, JULD_FIRST_MESSAGE <dbl>,
#> #   JULD_FIRST_MESSAGE_STATUS <chr>, JULD_FIRST_LOCATION <dbl>,
#> #   JULD_FIRST_LOCATION_STATUS <chr>, JULD_LAST_LOCATION <dbl>,
#> #   JULD_LAST_LOCATION_STATUS <chr>, JULD_LAST_MESSAGE <dbl>,
#> #   JULD_LAST_MESSAGE_STATUS <chr>, JULD_TRANSMISSION_END <dbl>,
#> #   JULD_TRANSMISSION_END_STATUS <chr>, CLOCK_OFFSET <dbl>, GROUNDED <chr>,
#> #   REPRESENTATIVE_PARK_PRESSURE <dbl>,
#> #   REPRESENTATIVE_PARK_PRESSURE_STATUS <chr>, CONFIG_MISSION_NUMBER <int>,
#> #   CYCLE_NUMBER_INDEX <int>, CYCLE_NUMBER_INDEX_ADJUSTED <int>,
#> #   DATA_MODE <chr>

argo_read_traj_history(traj_file)

#> # A tibble: 1 x 14
#>   float HISTORY_PREVIOU… HISTORY_INDEX_D… HISTORY_START_I… HISTORY_STOP_IN…
#>   <chr>            <dbl> <chr>                       <int>            <int>
#> 1 csio…               NA " "                            NA               NA
#> # … with 9 more variables: HISTORY_INSTITUTION <chr>, HISTORY_STEP <chr>,
#> #   HISTORY_SOFTWARE <chr>, HISTORY_SOFTWARE_RELEASE <chr>,
#> #   HISTORY_REFERENCE <chr>, HISTORY_DATE <chr>, HISTORY_ACTION <chr>,
#> #   HISTORY_PARAMETER <chr>, HISTORY_QCTEST <chr>